% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{adjust_color}
\alias{adjust_color}
\title{Correct duplicated colors}
\usage{
adjust_color(colorset)
}
\arguments{
\item{colorset}{a character vector of R or hex colors}
}
\value{
a character vector of hex colors with duplicated colors replaced
}
\description{
Correct duplicated colors
}
\examples{

original_colors <- c("#00FF00","#00FF00","#FF0000","#00FF00")

new_colors <- adjust_color(original_colors)

}
