% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{batch_process}
\alias{batch_process}
\title{Batch process}
\usage{
batch_process(x, batch.size, FUN, mc.cores = 1, .combine = "c", ...)
}
\arguments{
\item{x}{}

\item{batch.size}{}

\item{FUN}{}

\item{mc.cores}{}

\item{.combine}{}

\item{...}{}
}
\description{
Batch process
}
