% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{build_dend}
\alias{build_dend}
\title{Build dend}
\usage{
build_dend(
  cl.dat,
  cl.cor = NULL,
  l.rank = NULL,
  l.color = NULL,
  nboot = 100,
  ncores = 1
)
}
\arguments{
\item{cl.dat}{}

\item{cl.cor}{}

\item{l.rank}{}

\item{l.color}{}

\item{nboot}{}

\item{ncores}{}
}
\description{
Build dend
}
