% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{build_dend_harmonize}
\alias{build_dend_harmonize}
\title{Build dend harmonize}
\usage{
build_dend_harmonize(impute.dat.list, cl, cl.df, ncores = 1)
}
\arguments{
\item{impute.dat.list}{}

\item{cl}{}

\item{cl.df}{}

\item{ncores}{}
}
\value{

}
\description{
Build dend harmonize
}
