% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_tSNE.R
\name{build_tsne}
\alias{build_tsne}
\title{Title}
\usage{
build_tsne(
  dat,
  select.genes = rownames(dat),
  select.samples = colnames(dat),
  dims = 2,
  ...
)
}
\arguments{
\item{dat}{}

\item{select.genes}{}

\item{select.samples}{}

\item{dims}{}

\item{...}{}
}
\description{
Title
}
