% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calc_tau}
\alias{calc_tau}
\title{Calculate Tau scores for each gene}
\usage{
calc_tau(m, byRow = TRUE)
}
\arguments{
\item{m}{Matrix of expression values}

\item{byRow}{if TRUE, treats genes as row values and samples as columns.}
}
\value{
a vector of Tau scores
}
\description{
Calculate Tau scores for each gene
}
