% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions.R
\name{check_donor}
\alias{check_donor}
\title{Title}
\usage{
check_donor(anno, cluster, norm.dat, select.cells = names(cluster),
  keep.cl = NULL, meta1 = "roi", meta2 = "external_donor_name",
  plot = TRUE, plot.path = "output/")
}
\arguments{
\item{anno}{anno dataframe with the following required columns - facs_population_plan", "cluster"}

\item{cluster}{clusters labels for all cells along with sample_id as their names}

\item{norm.dat}{expression dataframe with columns as cells and rows as gene names and cpm normalized}

\item{select.cells}{column nmaes of norm.dat}

\item{keep.cl}{default is NULL}

\item{meta1}{default is "roi" (colname should be present in anno dataframe)}

\item{meta2}{default is "external_donor_name" (colname should be present in anno dataframe)}

\item{plot}{default is TRUE}

\item{plot.path}{path of plot, default is "./output/"}
}
\value{
Gives cluster ids vs donor expression/region heatmap
}
\description{
Title
}
\examples{
check_donor()
}
