% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions.R
\name{check_qc}
\alias{check_qc}
\title{Check qc of some numeric metrics}
\usage{
check_qc(x, qc.iqr.mult = 3)
}
\arguments{
\item{x}{numeric metric to check qc on}

\item{qc.iqr.mult}{default is 3}
}
\value{
returns binary result of whether qc failed or passed
}
\description{
Check qc of some numeric metrics
}
\examples{
check_qc(x,  qc.iqr.mult = 3)
}
