% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusCluster.R
\name{collect_co_matrix}
\alias{collect_co_matrix}
\title{Collect co-clustering matrix from results files}
\usage{
collect_co_matrix(result.files, all.cells)
}
\arguments{
\item{result.files}{A directory containing results files}

\item{all.cells}{The cells to read from results files}
}
\description{
Collect co-clustering matrix from results files
}
