% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusCluster.R
\name{collect_co_matrix_sparseM}
\alias{collect_co_matrix_sparseM}
\title{Collect coclustering results as a sparse matrix}
\usage{
collect_co_matrix_sparseM(norm.dat, result.files, all.cells,
  max.cl.size = 1000)
}
\arguments{
\item{norm.dat}{Normalized data matrix}

\item{result.files}{A directory containing results files}

\item{all.cells}{Samples to read from the results files}

\item{max.cl.size}{Maximum number of cells per cluster to use}
}
\description{
Collect coclustering results as a sparse matrix
}
