% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_util.R
\name{comb_de_result}
\alias{comb_de_result}
\title{Title}
\usage{
comb_de_result(
  de.genes.list,
  cl.means.list,
  common.genes = NULL,
  max.num = 10000,
  pairs = NULL,
  frac.th = 0.7
)
}
\arguments{
\item{de.genes.list}{}

\item{cl.means.list}{}

\item{common.genes}{}

\item{max.num}{}

\item{pairs}{}

\item{frac.th}{}
}
\description{
Title
}
