% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R, R/harmonize_util.R
\name{combine_cl}
\alias{combine_cl}
\title{combine cl}
\usage{
combine_cl(all.results)

combine_cl(all.results)
}
\arguments{
\item{all.results}{}
}
\value{



}
\description{
combine cl

Title
}
