% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions.R
\name{compare_plot}
\alias{compare_plot}
\title{Compare and plot two sets of cluster assignments}
\usage{
compare_plot(cl, ref.cl)
}
\arguments{
\item{cl}{A cluster factor object to compare to a reference}

\item{ref.cl}{A cluster factor object for the reference clusters}
}
\value{
g A ggplot2 dot plot object for the comparison.
}
\description{
This is the subset of the `compare_annotate` function that does the plotting.
}
