% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{compute_knn}
\alias{compute_knn}
\title{compute knn}
\usage{
compute_knn(
  comb.dat,
  select.genes,
  ref.list,
  select.sets = names(comb.dat$dat.list),
  select.cells = comb.dat$all.cells,
  k = 15,
  method = "cor",
  self.method = "RANN",
  batch.size = 10000,
  mc.cores = 1
)
}
\arguments{
\item{comb.dat}{}

\item{select.genes}{}

\item{ref.list}{}

\item{select.sets}{}

\item{select.cells}{}

\item{k}{}

\item{method}{}

\item{self.method}{}

\item{batch.size}{}

\item{mc.cores}{}
}
\description{
compute knn
}
