% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{compute_pair_deScore}
\alias{compute_pair_deScore}
\title{Compute deScore values based on a differential results data.frame and de_param().}
\usage{
compute_pair_deScore(df, de.param = de_param(), cl.size1 = NULL,
  cl.size2 = NULL)
}
\arguments{
\item{cl.size2}{}
}
\description{
Compute deScore values based on a differential results data.frame and de_param().
}
