% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vg.R
\name{compute_vg_stats}
\alias{compute_vg_stats}
\title{Compute gene expression variance statistics}
\usage{
compute_vg_stats(dat, verbose = FALSE)
}
\arguments{
\item{dat}{A matrix or dgCMatrix with samples as columns and genes as rows}

\item{verbose}{logical - whether or not to display status messages}
}
\value{
a data.frame with gene symbols, mean, variance, dispersion, z score, pvalues, and loess statistics
}
\description{
Compute gene expression variance statistics
}
