% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{convert_pair_matrix_str}
\alias{convert_pair_matrix_str}
\title{Convert paired cluster comparison values to a matrix}
\usage{
convert_pair_matrix_str(pair.str, l = NULL, directed = FALSE)
}
\arguments{
\item{l}{labels for columns. Default is NULL, which will compute them from names(pair.num)}

\item{directed}{If FALSE (default), the first value in each split will be used as columns, with the second as rows. 
If TRUE, first values will be rows, and second will be columns.}

\item{pair.num}{a vector of values with names of compared elements separted by "_", e.g. "c1_c2", "c23_c59"}
}
\value{
a matrix containing values from pair.num, and named for each element separated by "_" in names(pair.num)
}
\description{
Convert paired cluster comparison values to a matrix
}
\examples{

pair_values <- seq(1,27,3)
names(pair_values) <- paste(rep(letters[1:3], each = 3), rep(letters[1:3], 3), sep = "_")
pair_values

pair_matrix <- convert_pair_matrix(pair_values, directed = FALSE)
pair_matrix

pair_matrix <- convert_pair_matrix(pair_values, directed = TRUE)
pair_matrix

}
