% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{create_pairs}
\alias{create_pairs}
\title{Title}
\usage{
create_pairs(cn, direction = "nondirectional", include.self = FALSE)
}
\arguments{
\item{cn}{}

\item{direction}{}

\item{include.self}{}
}
\description{
Title
}
