% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{de_all_pairs}
\alias{de_all_pairs}
\title{Perform all pairwise differential expression comparison between clusters}
\usage{
de_all_pairs(norm.dat, cl, ...)
}
\arguments{
\item{...}{Additional parameters passed to DE_genes_pairs()}
}
\value{
a list containing DE results for every pair of clusters
}
\description{
Perform all pairwise differential expression comparison between clusters
}
\seealso{
\link{DE_genes_pairs}
}
