% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{DE_genes_cat_by_cl}
\alias{DE_genes_cat_by_cl}
\title{Title}
\usage{
DE_genes_cat_by_cl(norm.dat, cl, binary.cat, ...)
}
\arguments{
\item{norm.dat}{}

\item{cl}{}

\item{binary.cat}{}

\item{...}{}
}
\value{

}
\description{
Title
}
