% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_function_names.R
\name{DE_genes_pairs}
\alias{DE_genes_pairs}
\title{reverse-compatibility wrapper around de_selected_pairs()}
\usage{
DE_genes_pairs(...)
}
\arguments{
\item{...}{parameters passed to \code{link{de_selected_pairs}}}
}
\description{
reverse-compatibility wrapper around de_selected_pairs()
}
\seealso{
\code{link{de_selected_pairs}}
}
