% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_function_names.R
\name{DE_genes_pw}
\alias{DE_genes_pw}
\title{reverse-compatibility wrapper around de_all_pairs()}
\usage{
DE_genes_pw(...)
}
\description{
reverse-compatibility wrapper around de_all_pairs()
}
\seealso{
\code{link{de_all_pairs}}
}
