% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_function_names.R
\name{de_pair}
\alias{de_pair}
\title{reverse-compatibility wrapper around score_pair_stats()}
\usage{
de_pair(...)
}
\description{
reverse-compatibility wrapper around score_pair_stats()
}
\seealso{
\code{link{score_pair_stats}}
}
