% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{de_pair_limma}
\alias{de_pair_limma}
\title{Perform pairwise DE tests using limma for a single pair of clusters}
\usage{
de_pair_limma(pair, cl.present, cl.means, design, fit, genes)
}
\arguments{
\item{pair}{a numeric vector of length 2 specifying which clusters to compare}

\item{cl.present}{a data.frame of gene detection proportions (genes x clusters)}

\item{cl.means}{a data.frame of normalized mean gene expression values (genes x clusters)}

\item{design}{a limma design object}

\item{fit}{a limma fit object}

\item{genes}{the genes to use for pairwise comparisons}
}
\value{
a data.frame with DE statistics:
\itemize{
\item{padj} P-values adjusted using the Holm (1979) method (\code{p.adjust()} default).
\item{pval} P-values reported by the \code{limma::eBayes()} function.
\item{lfc} Log fold change of mean expression values between the pair of clusters.
\item{meanA} Normalized mean expression value for the first cluster in the pair.
\item{meanB} Normalized mean expression value for the second cluster in the pair.
\item{q1} Proportion of cells expressing each gene for the first cluster in the pair.
\item{q2} Proportion of cells expressing each gene for the second cluster in the pair.
}
}
\description{
Perform pairwise DE tests using limma for a single pair of clusters
}
