% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{de_param}
\alias{de_param}
\title{Set differential expression (DE) threshold for genes and clusters.}
\usage{
de_param(low.th = 1, padj.th = 0.01, lfc.th = 1, q1.th = 0.5,
  q2.th = NULL, q.diff.th = 0.7, de.score.th = 100, min.cells = 4)
}
\arguments{
\item{low.th}{Threshold on normalized gene expression to determiend whether a gene is detected in a cell. Default value 1. Users can specifiy different thresholds for different genes if necessary by using a vector with genes as names.}

\item{padj.th}{Threshold on adjusted Pvalue for limma. Default vaue is 0.01.}

\item{lfc.th}{Threshold on log2 fold change for limma. Default value is 1.}

\item{q1.th}{For every pair of clusters (one as foreground, and the other as background), we define q1, and q2 as the proportion of cells detecting the gene in the foregound and background cluster respectively. Up reguated genes should have q1 greater than q1.th}

\item{q2.th}{Up regualted genes should have q2 smaller than q2.th. q2.th is NULL by default.}

\item{q.diff.th}{Threshold on scaled proportion difference, defined as abs(q1 - q2)/max(q1, q2). DE genes should have differences greater than q.diff.th. Default 0.7}

\item{de.score.th}{To determine whether two clusters are seperable based on DE genes, we define de.score as the sum of -log10(adjusted Pvalub) for all DE genes. Each gene contributes at most 20 towards the sum. All clusters should have pairwise de.score greater than de.score.th.}
}
\value{
returns a data frame
}
\description{
Set differential expression (DE) threshold for genes and clusters.
}
