% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_function_names.R
\name{de_score_pairs}
\alias{de_score_pairs}
\title{reverse-compatibility wrapper around de_selected_pairs_stats()}
\usage{
de_score_pairs(...)
}
\description{
reverse-compatibility wrapper around de_selected_pairs_stats()
}
\seealso{
\code{link{de_selected_pairs_stats}}
}
