% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{de_selected_pairs}
\alias{de_selected_pairs}
\title{Perform pairwise differential gene expression tests between main pairs of clusters in parallel}
\usage{
de_selected_pairs(
  norm.dat,
  cl,
  pairs,
  method = "limma",
  low.th = 1,
  min.cells = 4,
  cl.means = NULL,
  cl.present = NULL,
  use.voom = FALSE,
  counts = NULL,
  mc.cores = 1
)
}
\arguments{
\item{norm.dat}{a normalized data matrix for data.}

\item{cl}{a cluster factor object.}

\item{pairs}{A 2-column matrix of cluster pairs.}

\item{method}{Either "limma" or "chisq".}

\item{low.th}{The minimum expression value used to filter for expressed genes.}

\item{cl.present}{A matrix of proportions of cells in each cluster with gene detection. Can be generated with \code{get_cl_props()}. Default is NULL (will be generated).}

\item{use.voom}{Logical, whether or not to use \code{voom()} for \code{limma} calculations. Default is FALSE.}

\item{counts}{A matrix of raw count data for each cell. Required if \code{use.voom} is TRUE. Default is NULL.}

\item{mc.cores}{A number indicating how many processor cores to use for parallelization.}
}
\value{

}
\description{
Perform pairwise differential gene expression tests between main pairs of clusters in parallel
}
