% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{de_stats_pair}
\alias{de_stats_pair}
\title{Compute differential expression summary statistics based on a differential results data.frame and de_param().}
\usage{
de_stats_pair(
  df,
  de.param = de_param(),
  cl.size1 = NULL,
  cl.size2 = NULL,
  select.genes = NULL
)
}
\arguments{
\item{df}{A data.frame of pairwise differential expression results (i.e. from \code{score_selected_pairs()}).}

\item{de.param}{A list of differential gene expression parameters from \code{de_param()}}

\item{cl.size1}{Optional: The number of samples in the first/high cluster}

\item{cl.size2}{Optional: The number of samples in the second/low cluster}
}
\description{
Compute differential expression summary statistics based on a differential results data.frame and de_param().
}
