% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{de_stats_selected_pairs}
\alias{de_stats_selected_pairs}
\title{Compute differential expression summary statistics for selected pairs of clusters based on de_param().}
\usage{
de_stats_selected_pairs(norm.dat, cl, pairs, de.df = NULL,
  de.param = de_param(), method = "limma", select.genes = NULL,
  mc.cores = 1, cl.means = NULL, cl.present = NULL)
}
\arguments{
\item{norm.dat}{a normalized data matrix for data.}

\item{cl}{a cluster factor object.}

\item{pairs}{A 2-column matrix of cluster pairs.}

\item{de.df}{Optional. Pre-computed results from \code{de_selected_pairs()} using the same pairs. Default = NULL.}

\item{de.param}{A list of differential gene expression parameters from \code{de_param()}}

\item{method}{If de.df is NULL, use "limma" or "chisq" to compute differentially expressed genes.}

\item{mc.cores}{If de.df is NULL, number of cores to use for parallel computation.}
}
\value{
A list with two objects:
\itemize{
\item{de.df} A list of results from \code{de_selected_pairs()} for each pair.
\item{de.genes} A list of results from \code{de_stats_pair()} for each pair.
}
}
\description{
Compute differential expression summary statistics for selected pairs of clusters based on de_param().
}
