% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{dend_lca}
\alias{dend_lca}
\title{Dend lca}
\usage{
dend_lca(dend, l1, l2, l = rep(attr(dend, "label"), length(l1)))
}
\arguments{
\item{dend}{}

\item{l1}{}

\item{l2}{}

\item{l}{}
}
\description{
Dend lca
}
