% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{display_cl}
\alias{display_cl}
\title{Display cluster plot}
\usage{
display_cl(
  cl,
  norm.dat,
  prefix = NULL,
  plot = !is.null(prefix),
  col = NULL,
  max.cl.size = NULL,
  markers = NULL,
  de.genes = NULL,
  main = "",
  height = 13,
  width = 9,
  min.sep = 10,
  ...
)
}
\arguments{
\item{cl}{}

\item{norm.dat}{}

\item{prefix}{}

\item{plot}{}

\item{col}{}

\item{max.cl.size}{}

\item{markers}{}

\item{de.genes}{}

\item{main}{}

\item{height}{}

\item{width}{}

\item{min.sep}{}

\item{...}{}
}
\description{
Display cluster plot
}
\author{
Zizhen Yao
}
