% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap4.R
\name{display_cl4}
\alias{display_cl4}
\title{Title}
\usage{
display_cl4(
  cl,
  norm.dat,
  prefix = NULL,
  plot = !is.null(prefix),
  col = NULL,
  max.cl.size = NULL,
  markers = NULL,
  de.genes = NULL,
  main = "",
  height = 13,
  width = 9,
  min.sep = 10,
  legend.dat = NULL,
  ...
)
}
\arguments{
\item{cl}{}

\item{norm.dat}{}

\item{prefix}{}

\item{plot}{}

\item{col}{}

\item{max.cl.size}{}

\item{markers}{}

\item{de.genes}{}

\item{main}{}

\item{height}{}

\item{width}{}

\item{min.sep}{}

\item{legend.dat}{}

\item{...}{}
}
\value{

}
\description{
Title
}
