% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{display_cl_one_vs_others}
\alias{display_cl_one_vs_others}
\title{Title}
\usage{
display_cl_one_vs_others(
  select.cl,
  cl,
  norm.dat,
  de.genes,
  plot = !is.null(prefix),
  col = NULL,
  max.cl.size = NULL,
  main = "",
  height = 13,
  width = 9,
  min.sep = 4,
  ...
)
}
\arguments{
\item{select.cl}{}

\item{cl}{}

\item{norm.dat}{}

\item{de.genes}{}

\item{plot}{}

\item{col}{}

\item{max.cl.size}{}

\item{main}{}

\item{height}{}

\item{width}{}

\item{min.sep}{}

\item{...}{}
}
\value{

}
\description{
Title
}
