% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN.graph.R
\name{edgeMaker}
\alias{edgeMaker}
\title{function to draw (curved) line between two points}
\usage{
edgeMaker(whichRow, len = 100, line.segments, curved = FALSE)
}
\arguments{
\item{whichRow}{}

\item{len}{}

\item{line.segments}{}

\item{curved}{}
}
\value{

}
\description{
function to draw (curved) line between two points
}
