% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_tSNE.R
\name{fast_tsne}
\alias{fast_tsne}
\title{Title}
\usage{
fast_tsne(
  dat,
  theta = 0.05,
  nthreads = 12,
  perplexity = 20,
  fast.tsne.path = "~/src/FIt-SNE",
  ...
)
}
\arguments{
\item{dat}{}

\item{theta}{}

\item{nthreads}{}

\item{perplexity}{}

\item{fast.tsne.path}{}

\item{...}{}
}
\description{
Title
}
