% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_noise_cl.R
\name{find_doublet}
\alias{find_doublet}
\title{Find doublet}
\usage{
find_doublet(cl.df, cl.sim, cl.good, de.genes = NULL)
}
\arguments{
\item{cl.df}{}

\item{cl.sim}{}

\item{cl.good}{}

\item{de.genes}{}
}
\value{

}
\description{
Find doublet
}
