% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_noise_cl.R
\name{find_low_quality_cl}
\alias{find_low_quality_cl}
\title{Use DEGenes to identify clusters with few differentially expressed genes}
\usage{
find_low_quality_cl(cl.df, cl.good, de.score.mat = NULL, de.genes)
}
\arguments{
\item{cl.df}{A data.frame with cluster annotations that includes a "cluster_label" column}

\item{cl.good}{a cluster factor object for cells from high-quality clusters.}

\item{de.genes}{differential gene expression results like those generated by de_score().}
}
\value{
a data.frame with the closest "good" cluster for each non-good/low cluster.
}
\description{
To use this, you'll need to perform pairwise gene expression tests with de_score(), and 
a set of clusters with high confidence (cl.good).
}
\details{
Low-quality clusters will be from all clusters in cl.df that are not in cl.good.
}
