% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vg.R
\name{find_vg}
\alias{find_vg}
\title{Identify high variability genes using Brennecke's method}
\usage{
find_vg(dat, plot_file = NULL, return_type = "data")
}
\arguments{
\item{dat}{A matrix or dgCMatrix with samples as columns and genes as rows}

\item{plot_file}{(optional) A file to use to output diagnostic plots. Default is NULL.}

\item{return_type}{What kind of objects to return. Default is "data", which will return only 
a data.frame with results for each gene. Other options are "plots", which will return a list 
containing plot objects, and "both" which will return a list with the statistics (gene_var_stats) 
and plots.}
}
\value{
Either a data.frame or a list (see return_type parameter).
}
\description{
This function is a wrapper around compute_vg_stats(), which computes variability statistics, 
and plot_vg(), which generates multiple plots for those results.
}
