% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vg.R
\name{gene_dispersion}
\alias{gene_dispersion}
\title{Compute dispersion for each gene in a gene x sample matrix}
\usage{
gene_dispersion(dat, means = NULL, vars = NULL, rescale = FALSE)
}
\arguments{
\item{dat}{a matrix or dg[CT]Matrix of expression values with genes as rows and samples as columns.}

\item{means}{(optional) a vector of means for each gene/row. If NULL (default), will be computed using gene_means().}

\item{vars}{(optional) a vector of variances for each gene/row. If NULL (default), will be computed using gene_vars().}

\item{rescale}{a logical indicating whether or not to rescale using rescale_samples(). Default is FALSE.}
}
\value{
a numeric vector of dispersion values for each gene
}
\description{
Compute dispersion for each gene in a gene x sample matrix
}
