% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_util.R
\name{gene_gene_cor_conservation}
\alias{gene_gene_cor_conservation}
\title{Title}
\usage{
gene_gene_cor_conservation(dat.list, select.genes, select.cells, pairs = NULL)
}
\arguments{
\item{dat.list}{}

\item{select.genes}{}

\item{select.cells}{}

\item{pairs}{}
}
\description{
Title
}
