% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vg.R
\name{gene_loess_fit}
\alias{gene_loess_fit}
\title{Compute dispersion-mean Loess-fit for each gene in a gene x sample matrix}
\usage{
gene_loess_fit(dat, dispersions = NULL, means = NULL, rescale = FALSE)
}
\arguments{
\item{dat}{a matrix or dg[CT]Matrix of expression values with genes as rows and samples as columns.}

\item{dispersions}{(optional) a vector of dispersions for each gene/row. If NULL (default), will be computed using gene_dispersion().}

\item{means}{(optional) a vector of means for each gene/row. If NULL (default), will be computed using gene_means().}

\item{rescale}{a logical indicating whether or not to rescale using rescale_samples(). Default is FALSE.}
}
\value{
a loess fit object returned by the loess() function.
}
\description{
Compute dispersion-mean Loess-fit for each gene in a gene x sample matrix
}
