% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vg.R
\name{gene_loess_fit_z}
\alias{gene_loess_fit_z}
\title{Compute dispersion z scores for the residuals of a loess fit}
\usage{
gene_loess_fit_z(loess_fit, dispersions)
}
\arguments{
\item{loess_fit}{an object output by the loess() function, with class "loess".}

\item{dispersions}{gene dispersion values as generated by the gene_dispersion() function.}
}
\value{
a numeric vector of z score values
}
\description{
Compute dispersion z scores for the residuals of a loess fit
}
