% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vg.R
\name{gene_vars}
\alias{gene_vars}
\title{Compute variance for each gene in a gene x sample matrix}
\usage{
gene_vars(dat, means = NULL, rescale = FALSE)
}
\arguments{
\item{dat}{a matrix or dg[CT]Matrix of expression values with genes as rows and samples as columns.}

\item{means}{(optional) a vector of means for each gene/row. If NULL (default), will be computed using gene_means().}

\item{rescale}{a logical indicating whether or not to rescale using rescale_samples(). Default is FALSE.}
}
\value{
a numeric vector of variance values for each gene
}
\description{
This function uses mean(values ^ 2) - mean(values) ^ 2
}
\details{
Results will differ from the var() function.
}
