% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vg.R
\name{gene_z}
\alias{gene_z}
\title{Compute dispersion z-scores for each gene in a gene x sample matrix}
\usage{
gene_z(dat, dispersions = NULL, rescale = FALSE)
}
\arguments{
\item{dat}{a matrix or dg[CT]Matrix of expression values with genes as rows and samples as columns.}

\item{dispersions}{(optional) a vector of dispersions for each gene/row. If NULL (default), will be computed using gene_dispersion().}

\item{rescale}{a logical indicating whether or not to rescale using rescale_samples(). Default is FALSE.}
}
\value{
a numeric vector of z-scores for each gene
}
\description{
This version of z scoring will differ from base R's scale()
}
