% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{get_beta_score}
\alias{get_beta_score}
\title{Title}
\usage{
get_beta_score(propExpr, spec.exp = 2, mcores = 1)
}
\arguments{
\item{propExpr}{}

\item{spec.exp}{}

\item{mcores}{}
}
\value{

}
\description{
Title
}
