% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusCluster.R
\name{get_cell.cl.co.ratio}
\alias{get_cell.cl.co.ratio}
\title{Get cell co-clustering ratios}
\usage{
get_cell.cl.co.ratio(cl, co.ratio = NULL, cl.mat = NULL)
}
\arguments{
\item{cl}{Vector of cluster assignments}

\item{co.ratio}{coclustering ratio results}

\item{cl.mat}{Cluster membership matrix for all cells and all clusters from all bootstrapping iterations.}
}
\description{
Get cell co-clustering ratios
}
