% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{get_cl_df}
\alias{get_cl_df}
\title{Generate an initial cl.df object based on cl}
\usage{
get_cl_df(cl)
}
\arguments{
\item{cl}{a cluster factor object}
}
\value{
a data.frame with an id, color, and size for each cluster.
}
\description{
Generate an initial cl.df object based on cl
}
