% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_cl_mat}
\alias{get_cl_mat}
\title{Generate a sparse matrix one-hot representation of clusters x samples}
\usage{
get_cl_mat(cl)
}
\arguments{
\item{cl}{a cluster factor object}
}
\value{
a sparse, one-hot matrix indicating which cluster(columns) each sample (rows) belongs to.
}
\description{
Generate a sparse matrix one-hot representation of clusters x samples
}
