% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_cl_medians}
\alias{get_cl_medians}
\title{Compute cluster medians for each row in a matrix}
\usage{
get_cl_medians(mat, cl)
}
\arguments{
\item{mat}{A gene (rows) x samples (columns) sparse matrix}

\item{cl}{A cluster factor object}
}
\value{
a matrix of genes (rows) x clusters (columns) with medians for each cluster
}
\description{
Compute cluster medians for each row in a matrix
}
