% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_util.R
\name{get_cl_present_list}
\alias{get_cl_present_list}
\title{Title}
\usage{
get_cl_present_list(
  dat.list,
  de.param.list,
  select.genes = NULL,
  cl,
  sets = names(dat.list)
)
}
\arguments{
\item{dat.list}{}

\item{de.param.list}{}

\item{select.genes}{}

\item{cl}{}

\item{sets}{}
}
\value{

}
\description{
Title
}
