% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_cl_prop}
\alias{get_cl_prop}
\title{Compute cluster proportions for each row in a matrix}
\usage{
get_cl_prop(mat, cl, thresshold = 1)
}
\arguments{
\item{mat}{a gene (rows) x samples(columns) sparse matrix}

\item{cl}{A cluster factor object}

\item{thresshold}{The minimum expression value used to binarize the results}
}
\value{
a matrix of genes (rows) x cluster(columns) with proportions for each cluster
}
\description{
Compute cluster proportions for each row in a matrix
}
