% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_merge.R
\name{get_cl_sim_multiple}
\alias{get_cl_sim_multiple}
\title{Title}
\usage{
get_cl_sim_multiple(cl.rd.list, FUN = pmax)
}
\arguments{
\item{cl.rd.list}{}

\item{FUN}{}
}
\value{

}
\description{
Title
}
