% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_cl_sums}
\alias{get_cl_sums}
\title{Compute cluster sums for each row in a matrix}
\usage{
get_cl_sums(mat, cl)
}
\arguments{
\item{mat}{A gene (rows) x samples (columns) sparse matrix}

\item{cl}{A cluster factor object}
}
\value{
a matrix of genes (rows) x clusters (columns) with sums for each cluster
}
\description{
Compute cluster sums for each row in a matrix
}
