% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_cv.R
\name{get_core_intermediate}
\alias{get_core_intermediate}
\title{Get core intermediate}
\usage{
get_core_intermediate(
  norm.dat,
  cl,
  select.markers,
  n.bin = 5,
  n.iter = 100,
  mc.cores = 10,
  method = "median"
)
}
\arguments{
\item{norm.dat}{}

\item{cl}{}

\item{select.markers}{}

\item{n.bin}{}

\item{n.iter}{}

\item{mc.cores}{}

\item{method}{}
}
\value{

}
\description{
Get core intermediate
}
