% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{get_de_matrix}
\alias{get_de_matrix}
\title{Generate a matrix of pairwise DE results}
\usage{
get_de_matrix(de.genes, directed = FALSE, field = "num")
}
\arguments{
\item{de.genes}{Output from \code{de_score} or \code{de_stats_all_pairs}.}

\item{directed}{Logical indicating whether to select results based on all DE genes (Default, FALSE) or up/down regulated genes (see Details).}

\item{field}{The result to retrieve from de.results. Either "score" or "num". Default is "num".}
}
\value{
a matrix with clusters as rows and columns, and pairwise DE results as values.
}
\description{
Generate a matrix of pairwise DE results
}
\details{
When directed = TRUE and field = "num", the minimum value from up or down-regulated genes is returned for each pair. When field = "score", the 
minimum deScore is returned.
}
