% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_util.R
\name{get_de_result}
\alias{get_de_result}
\title{Title}
\usage{
get_de_result(
  dat.list,
  de.param.list,
  cl,
  select.sets = names(de.param.list),
  ...
)
}
\arguments{
\item{dat.list}{}

\item{de.param.list}{}

\item{cl}{}

\item{select.sets}{}

\item{...}{}
}
\description{
Title
}
