% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_noise_cl.R
\name{get_de_score}
\alias{get_de_score}
\title{Get de score}
\usage{
get_de_score(de.df, top.genes, upper = 20)
}
\arguments{
\item{de.df}{}

\item{top.genes}{}

\item{upper}{}
}
\value{

}
\description{
Get de score
}
