% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{get_dend_markers}
\alias{get_dend_markers}
\title{get dendrogram precomputed markers}
\usage{
get_dend_markers(dend, n = 20)
}
\arguments{
\item{dend}{}

\item{n}{}
}
\value{

}
\description{
get dendrogram precomputed markers
}
