% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{get_dend_markers_direction}
\alias{get_dend_markers_direction}
\title{get dendrogram precomputed markers (direction)}
\usage{
get_dend_markers_direction(dend, n = 20)
}
\arguments{
\item{dend}{}

\item{n}{}
}
\description{
get dendrogram precomputed markers (direction)
}
