% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_WGCNA.R
\name{get_eigen}
\alias{get_eigen}
\title{Compute module eigen genes}
\usage{
get_eigen(
  gene.mod,
  norm.dat,
  select.cells = colnames(norm.dat),
  prefix = NULL,
  method = "ward.D",
  hc = NULL,
  ...
)
}
\arguments{
\item{gene.mod}{A list of gene modules.}

\item{norm.dat}{log transformed normalized data matrix.}

\item{select.cells}{Cells used to compute module eigen genes.}

\item{prefix}{Default NULL. If not NULL, a heatmap of the gene module eigen genes will be produced with "prefix" as the prefix for the pdf file.}

\item{method}{Default "ward.D". Used by hclust method to create the cell dendrogram for the heatmap display.}

\item{hc}{Precomputed cell dendrogram for heatmap display. Default NULL.}

\item{...}{Other plotting parameters passed to the heatmap function.}
}
\value{
A list with two elements: module eigen genes, and if prefix is not NULL, dendrogram for selected cells.
}
\description{
Compute module eigen genes
}
