% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{get_gene_score}
\alias{get_gene_score}
\title{Title}
\usage{
get_gene_score(de.genes, top.n = 50, max.num = 1000, bin.th = 4)
}
\arguments{
\item{de.genes}{}

\item{top.n}{}

\item{max.num}{}

\item{bin.th}{}
}
\value{

}
\description{
Title
}
