% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{get_knn_batch}
\alias{get_knn_batch}
\title{get knn batch}
\usage{
get_knn_batch(
  dat,
  ref.dat,
  k,
  method = "cor",
  dim = NULL,
  batch.size,
  mc.cores = 1
)
}
\arguments{
\item{dat}{}

\item{ref.dat}{}

\item{k}{}

\item{method}{}

\item{dim}{}

\item{batch.size}{}

\item{mc.cores}{}
}
\description{
get knn batch
}
