% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNN.graph.R
\name{get_knn_graph}
\alias{get_knn_graph}
\title{get knn graph}
\usage{
get_knn_graph(
  rd.dat,
  cl,
  cl.df,
  k = 15,
  knn.outlier.th = 2,
  outlier.frac.th = 0.5
)
}
\arguments{
\item{rd.dat}{}

\item{cl}{}

\item{cl.df}{}

\item{k}{}

\item{knn.outlier.th}{}

\item{outlier.frac.th}{}
}
\description{
get knn graph
}
