% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_impute.R
\name{get_knn_weight}
\alias{get_knn_weight}
\title{Title}
\usage{
get_knn_weight(knn.dist, scale = 0.2, exclude.th = 1e-04)
}
\arguments{
\item{knn.dist}{}

\item{scale}{}

\item{exclude.th}{}
}
\value{

}
\description{
Title
}
