% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_pair_matrix}
\alias{get_pair_matrix}
\title{Subset a matrix as a vector using row and column positions}
\usage{
get_pair_matrix(m, rows, cols)
}
\arguments{
\item{m}{a matrix object}

\item{rows}{row positions, either as indices or matches to row names}

\item{cols}{row positions, either as indices or matches to row names}
}
\value{
a vector with values extracted from m at [rows/cols]
}
\description{
Subset a matrix as a vector using row and column positions
}
