% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_pair_matrix_coor}
\alias{get_pair_matrix_coor}
\title{Convert matrix row/column positions to vector position}
\usage{
get_pair_matrix_coor(m, rows, cols)
}
\arguments{
\item{m}{a matrix object}

\item{rows}{row positions, either as indices or matches to row names}

\item{cols}{row positions, either as indices or matches to row names}
}
\value{
a vector containing the numeric position at [rows,cols]
}
\description{
Convert matrix row/column positions to vector position
}
