% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_pairs}
\alias{get_pairs}
\title{Convert underscore_separated pair names to a data.frame}
\usage{
get_pairs(pairs.str)
}
\arguments{
\item{pairs.str}{a character vector with underscore_separated values}
}
\value{
a data.frame with columns "P1" and "P2" containing the separated pair names
}
\description{
Convert underscore_separated pair names to a data.frame
}
