% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions.R
\name{group_cl}
\alias{group_cl}
\title{Assign clusters to a group}
\usage{
group_cl(anno, cluster, norm.dat, select.cells = colnames(norm.dat),
  neun.thresh = 0.5, neun.colname = "facs_population_plan",
  neun.val = "NeuN-pos", test.genes = c("GAD1", "GAD2", "SLC17A7",
  "SLC17A6"), expr.th = 3, prop.th = c(0.5, 0.5, 0.2, 0.2),
  outlier.cl = NULL, donor.cl = NULL)
}
\arguments{
\item{anno}{anno dataframe which must include column names listed in `neun.colname`
"cluster" is added from `cluster` parameter below.}

\item{cluster}{clusters labels for all cells along with sample_id as their names}

\item{norm.dat}{expression dataframe with columns as cells and rows as gene names 
and cpm normalized}

\item{select.cells}{column names of norm.dat}

\item{neun.thresh}{fraction of cells expressing NeuN to be considered 
NeuN positive (default is 0.5)}

\item{neun.colname}{column name in anno with the Neun information}

\item{neun.val}{value corresponding to non-neuronal marker in neun.colname in anno}

\item{test.genes}{genes used to define cell classes (default is GAD1, GAD2, SLC17A7, SLC17A6)}

\item{expr.th}{expression threshold for detecting test genes}

\item{prop.th}{proportion threshold of detected genes by cluster (default is 0.5, 0.5, 0.2, 0.2)}

\item{outlier.cl}{vector of clusters previously defined as "outlier" (default is NULL)}

\item{donor.cl}{vector of clusters previously defined as "donor" (default is NULL)}
}
\value{
gives the clusters names in each class (exc, inh, glia, donor, outlier)
  as a list
}
\description{
This function assigns each cluster to a predefined set of classes (exc, inh, glia, 
  donor, outlier) using test.genes (default GAD1, GAD2, SLC17A7, SLC17A6).
  Later iterations of the function will allow for user-defined classes and genes.
}
