% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{group_specific_markers}
\alias{group_specific_markers}
\title{Title}
\usage{
group_specific_markers(
  cl.g,
  norm.dat,
  cl,
  de.param,
  n.markers = 5,
  cl.present.counts = NULL
)
}
\arguments{
\item{cl.g}{}

\item{norm.dat}{}

\item{cl}{}

\item{de.param}{}

\item{n.markers}{}

\item{cl.present.counts}{}
}
\value{

}
\description{
Title
}
