% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{harmonize}
\alias{harmonize}
\title{Harmonize}
\usage{
harmonize(comb.dat, prefix, overwrite = TRUE, dir = ".", ...)
}
\arguments{
\item{comb.dat}{}

\item{prefix}{}

\item{overwrite}{}

\item{dir}{}

\item{...}{}
}
\description{
Harmonize
}
