% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap.3}
\alias{heatmap.3}
\title{A modified version of heatmap.2 from the gplots package for cluster plotting}
\usage{
heatmap.3(x, Rowv = TRUE, Colv = if (symm) "Rowv" else TRUE,
  distfun = dist, hclustfun = hclust, dendrogram = c("both", "row",
  "column", "none"), symm = FALSE, scale = c("none", "row", "column"),
  na.rm = TRUE, revC = identical(Colv, "Rowv"), add.expr, breaks,
  symbreaks = min(x < 0, na.rm = TRUE) || scale != "none",
  col = "heat.colors", colsep, rowsep, sepcolor = "white",
  sepwidth = c(0.05, 0.05), cellnote, notecex = 1, notecol = "cyan",
  na.color = par("bg"), trace = c("column", "row", "both", "none"),
  tracecol = "cyan", hline = median(breaks), vline = median(breaks),
  linecol = tracecol, margins = c(5, 5), ColSideColors, RowSideColors,
  cexRow = 0.2 + 1/log10(nr), cexCol = 0.2 + 1/log10(nc),
  labRow = NULL, labCol = NULL, key = TRUE, keysize = 1.5,
  density.info = c("histogram", "density", "none"), denscol = tracecol,
  symkey = min(x < 0, na.rm = TRUE) || symbreaks, densadj = 0.25,
  main = NULL, xlab = NULL, ylab = NULL, lmat = NULL,
  lhei = NULL, lwid = NULL, ...)
}
\arguments{
\item{x}{numeric matrix of the values to be plotted.}

\item{Rowv}{determines if and how the \emph{row} dendrogram should be
reordered.  By default, it is TRUE, which implies dendrogram is
computed and reordered based on row means. If NULL or FALSE, then no
dendrogram is computed and no reordering is done. If a
\code{\link{dendrogram}}, then it is used "as-is", ie
without any reordering. If a vector of integers, then dendrogram is
computed and reordered based on the order of the vector.}

\item{Colv}{determines if and how the \emph{column} dendrogram should
be reordered. Has the options as the \code{Rowv} argument above and
\emph{additionally} when \code{x} is a square matrix,
\code{Colv="Rowv"} means that columns should be treated identically
to the rows.}

\item{distfun}{function used to compute the distance (dissimilarity)
between both rows and columns.  Defaults to \code{\link{dist}}.}

\item{hclustfun}{function used to compute the hierarchical clustering
when \code{Rowv} or \code{Colv} are not dendrograms.  Defaults to
\code{\link{hclust}}.}

\item{dendrogram}{character string indicating whether to draw 'none',
'row', 'column' or 'both' dendrograms.  Defaults to 'both'. However,
if Rowv (or Colv) is FALSE or NULL and dendrogram is 'both', then a
warning is issued and Rowv (or Colv) arguments are honoured.}

\item{symm}{logical indicating if \code{x} should be treated
\bold{symm}etrically; can only be true when \code{x} is a
square matrix.}

\item{scale}{character indicating if the values should be centered and
scaled in either the row direction or the column direction, or
none.  The default is \code{"none"}.}

\item{na.rm}{logical indicating whether \code{NA}'s should be removed.}

\item{revC}{logical indicating if the column order should be
\code{\link{rev}}ersed for plotting, such that e.g., for the
symmetric case, the symmetry axis is as usual.}

\item{add.expr}{expression that will be evaluated after the call to
\code{image}.  Can be used to add components to the plot.}

\item{breaks}{(optional) Either a numeric vector indicating the
splitting points for binning \code{x} into colors, or a integer
number of break points to be used, in which case the break points
will be spaced equally between \code{min(x)} and \code{max(x)}.}

\item{symbreaks}{Boolean indicating whether breaks should be
made symmetric about 0. Defaults to \code{TRUE} if the data includes
negative values, and to \code{FALSE} otherwise.}

\item{col}{colors used for the image. Defaults to heat colors
(\code{heat.colors}).}

\item{colsep, rowsep, sepcolor}{(optional) vector of integers
indicating which columns or rows should be separated from the
preceding columns or rows by a narrow space of color
\code{sepcolor}.}

\item{sepwidth}{(optional) Vector of length 2 giving the width
(colsep) or height (rowsep) the separator box drawn by colsep and
rowsep as a function of the width (colsep) or height (rowsep) of a
cell. Defaults to \code{c(0.05, 0.05)}}

\item{cellnote}{(optional) matrix of character strings which will be
placed within each color cell, e.g. p-value symbols.}

\item{notecex}{(optional) numeric scaling factor for \code{cellnote}
items.}

\item{notecol}{(optional) character string specifying the color for
\code{cellnote} text.  Defaults to "cyan".}

\item{na.color}{Color to use for missing value (\code{NA}). Defaults
to the plot background color.}

\item{trace}{character string indicating whether a solid "trace" line
should be drawn across 'row's or down 'column's, 'both' or 'none'.
The distance of the line from the center of each color-cell is
proportional to the size of the measurement. Defaults to 'column'.}

\item{tracecol}{character string giving the color for "trace"
line. Defaults to "cyan".}

\item{hline, vline, linecol}{Vector of values within cells where a
horizontal or vertical dotted line should be drawn.  The color of
the line is controlled by \code{linecol}.  Horizontal  lines are only
plotted if \code{trace} is 'row' or 'both'.  Vertical lines are only
drawn if \code{trace} 'column' or 'both'.   \code{hline} and
\code{vline} default to the median of the breaks, \code{linecol}
defaults to the value of \code{tracecol}.}

\item{margins}{numeric vector of length 2 containing the margins
(see \code{\link{par}(mar= *)}) for column and row names,
respectively.}

\item{ColSideColors}{(optional) character vector of length
\code{ncol(x)} containing the color names for a horizontal side bar
that may be used to annotate the columns of \code{x}.}

\item{RowSideColors}{(optional) character vector of length
\code{nrow(x)} containing the color names for a vertical side bar
that may be used to annotate the rows of \code{x}.}

\item{cexRow, cexCol}{positive numbers, used as \code{cex.axis} in
for the row or column axis labeling.  The defaults currently only
use number of rows or columns, respectively.}

\item{labRow, labCol}{character vectors with row and column labels to
use; these default to \code{rownames(x)} or \code{colnames(x)},
respectively.}

\item{keysize}{numeric value indicating the size of the key}

\item{density.info}{character string indicating whether to superimpose
a 'histogram', a 'density' plot, or no plot ('none') on the
color-key.}

\item{denscol}{character string giving the color for the density
display specified by \code{density.info}, defaults to the same value
as \code{tracecol}.}

\item{symkey}{Boolean indicating whether the color key should be
made symmetric about 0. Defaults to \code{TRUE} if the data includes
negative values, and to \code{FALSE} otherwise.}

\item{densadj}{Numeric scaling value for tuning the kernel width when
a density plot is drawn on the color key.  (See the \code{adjust}
parameter for the \code{density} function for details.)  Defaults to
0.25.}

\item{main, xlab, ylab}{main, x- and y-axis titles; defaults to none.}

\item{lmat, lhei, lwid}{visual layout: position matrix, column height,
column width.  See \code{\link{gplots::heatmap.2}} for details}

\item{...}{additional arguments passed on to \code{\link{image}}}

\item{reorderfun}{\code{function(d, w)} of dendrogram and weights for
reordering the row and column dendrograms.  The default uses
\code{\link{stats reorder.dendrogram}}.}

\item{srtRow, srtCol}{angle of row/column labels, in degrees from
horizontal}

\item{adjRow, adjCol}{2-element vector giving the (left-right,
top-bottom) justification of row/column labels (relative to the text
orientation).}

\item{offsetRow, offsetCol}{Number of character-width spaces to
place between row/column labels and the edge of the plotting
region.}

\item{colRow, colCol}{color of row/column labels, either a scalar to
set the color of all labels the same, or a vector providing the
colors of each label item}

\item{key.title}{main title of the color key. If set to NA no title
will be plotted.}

\item{key.xlab}{x axis label of the color key. If set to NA no label
will be plotted.}

\item{key.ylab}{y axis label of the color key. If set to NA no label
will be plotted.}

\item{key.xtickfun}{function computing tick location and labels for
the xaxis of the color key. Returns a named list containing
parameters that can be passed to \code{axis}. See examples.}

\item{key.ytickfun}{function computing tick location and labels for
the y axis of the color key. Returns a named list containing
parameters that can be passed to \code{axis}.  See examples.}

\item{key.par}{graphical parameters for the color key. Named list that
can be passed to \code{par}.}

\item{extrafun}{A function to be called after all other work. See
code{\link{gplots::heatmap.2}}.}
}
\value{
a base R plot
}
\description{
A modified version of heatmap.2 from the gplots package for cluster plotting
}
