% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap4.R
\name{heatmap.4}
\alias{heatmap.4}
\title{Title}
\usage{
heatmap.4(
  x,
  Rowv = TRUE,
  Colv = if (symm) "Rowv" else TRUE,
  distfun = dist,
  hclustfun = hclust,
  dendrogram = c("both", "row", "column", "none"),
  symm = FALSE,
  scale = c("none", "row", "column"),
  na.rm = TRUE,
  revC = identical(Colv, "Rowv"),
  add.expr,
  breaks,
  symbreaks = min(x < 0, na.rm = TRUE) || scale != "none",
  col = "heat.colors",
  colsep,
  rowsep,
  sepcolor = "white",
  sepwidth = c(0.05, 0.05),
  cellnote,
  notecex = 1,
  notecol = "cyan",
  na.color = par("bg"),
  trace = c("column", "row", "both", "none"),
  tracecol = "cyan",
  hline = median(breaks),
  vline = median(breaks),
  linecol = tracecol,
  margins = c(5, 5),
  ColSideColors,
  RowSideColors,
  cexRow = 0.2 + 1/log10(nr),
  cexCol = 0.2 + 1/log10(nc),
  labRow = NULL,
  labCol = NULL,
  key = TRUE,
  keysize = 1,
  density.info = c("histogram", "density", "none"),
  denscol = tracecol,
  symkey = min(x < 0, na.rm = TRUE) || symbreaks,
  densadj = 0.25,
  main = NULL,
  xlab = NULL,
  ylab = NULL,
  lmat = NULL,
  lhei = 0.01,
  lwid = NULL,
  ...
)
}
\arguments{
\item{x}{}

\item{Rowv}{}

\item{Colv}{}

\item{distfun}{}

\item{hclustfun}{}

\item{dendrogram}{}

\item{symm}{}

\item{scale}{}

\item{na.rm}{}

\item{revC}{}

\item{add.expr}{}

\item{breaks}{}

\item{symbreaks}{}

\item{col}{}

\item{colsep}{}

\item{rowsep}{}

\item{sepcolor}{}

\item{sepwidth}{}

\item{cellnote}{}

\item{notecex}{}

\item{notecol}{}

\item{na.color}{}

\item{trace}{}

\item{tracecol}{}

\item{hline}{}

\item{vline}{}

\item{linecol}{}

\item{margins}{}

\item{ColSideColors}{}

\item{RowSideColors}{}

\item{cexRow}{}

\item{cexCol}{}

\item{labRow}{}

\item{labCol}{}

\item{key}{}

\item{keysize}{}

\item{density.info}{}

\item{denscol}{}

\item{symkey}{}

\item{densadj}{}

\item{main}{}

\item{xlab}{}

\item{ylab}{}

\item{lmat}{}

\item{lhei}{}

\item{lwid}{}

\item{...}{}
}
\description{
Title
}
