% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{i_harmonize}
\alias{i_harmonize}
\title{i harmonize}
\usage{
i_harmonize(
  comb.dat,
  select.cells = comb.dat$all.cells,
  ref.sets = names(comb.dat$dat.list),
  prefix = "",
  result = NULL,
  overwrite = TRUE,
  ...
)
}
\arguments{
\item{comb.dat}{}

\item{select.cells}{}

\item{ref.sets}{}

\item{prefix}{}

\item{result}{}

\item{overwrite}{}

\item{...}{}
}
\value{

}
\description{
i harmonize
}
