% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_impute.R
\name{impute_knn_global}
\alias{impute_knn_global}
\title{Title}
\usage{
impute_knn_global(
  comb.dat,
  split.results,
  select.genes,
  select.cells,
  ref.list,
  sets = comb.dat$sets,
  max.dim = 80,
  th = 0.5,
  rm.eigen = NULL,
  rm.th = 0.65
)
}
\arguments{
\item{comb.dat}{}

\item{split.results}{}

\item{select.genes}{}

\item{select.cells}{}

\item{ref.list}{}

\item{sets}{}

\item{max.dim}{}

\item{th}{}

\item{rm.eigen}{}

\item{rm.th}{}
}
\description{
Title
}
