% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R, R/harmonize_util.R
\name{impute_val_cor}
\alias{impute_val_cor}
\title{Impute val cor}
\usage{
impute_val_cor(dat, impute.dat)

impute_val_cor(dat, impute.dat)
}
\arguments{
\item{dat}{}

\item{impute.dat}{}
}
\description{
Impute val cor

Title
}
