% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions_rename.r
\name{incrementHex}
\alias{incrementHex}
\title{Adjust color vector}
\usage{
incrementHex(col, r = 0, g = 0, b = 0)
}
\arguments{
\item{col}{a vector of colors (in a format compatable with col2rgb)}

\item{r}{integer value to increment red (-255 to 255)}

\item{g}{integer value to increment green (-255 to 255)}

\item{b}{integer value to increment blue (-255 to 255)}
}
\value{
a vector with all colors adjusted by the r, g, b integer 
values and in #xxxxxx format.
}
\description{
Note: This function is an internal function
}
