% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{iter_clust}
\alias{iter_clust}
\title{Iterative clustering algorithm for single cell RNAseq dataset}
\usage{
iter_clust(norm.dat, select.cells = colnames(norm.dat), prefix = NULL,
  split.size = 10, result = NULL, method = "auto", ...)
}
\arguments{
\item{norm.dat}{normalized expression data matrix in log transform, using genes as rows, and cells and columns. Users can use log2(FPKM+1) or log2(CPM+1)}

\item{select.cells}{The cells to be clustered}

\item{prefix}{The character string to indicate current iteration.}

\item{split.size}{The minimal cluster size for further splitting}

\item{result}{The current clustering result as basis for further splitting.}

\item{method}{Clustering method. It can be "auto", "louvain", "hclust"}

\item{...}{Other parameters passed to method `onestep_clust()`}
}
\value{
Clustering result is returned as a list with two elements: 
        cl: cluster membership for each cell
        markers: top markers that seperate clusters
}
\description{
Iterative clustering algorithm for single cell RNAseq dataset
}
\examples{
clust.result <- iter_clust(norm.dat)
          clust.result <- iter_clust(norm.dat, de.param = de_param(q1.th = 0.5, de.score.th = 100))
}
