% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusCluster.R
\name{iter_consensus_clust}
\alias{iter_consensus_clust}
\title{Iterative consensus clustering}
\usage{
iter_consensus_clust(co.ratio, cl.list, norm.dat,
  select.cells = colnames(co.ratio), all.col = NULL, diff.th = 0.25,
  prefix = NULL, method = c("auto", "louvain", "ward.D"), verbose = FALSE,
  de.param = de.param, result = NULL, rd.dat = NULL)
}
\arguments{
\item{co.ratio}{cell cell co-clustering matrix}

\item{cl.list}{The list of subsampled clustering results.}

\item{norm.dat}{The log2 transformed normalzied expression matrix}

\item{select.cells}{Cells to be clustered}

\item{all.col}{Color bars for plotting heatmap. Default NULL}

\item{diff.th}{The difference of co-clustering probablities for splitting a cluster.}

\item{prefix}{Default NULL.}

\item{method}{Clustering methods. Default "auto"}

\item{verbose}{Default FALSE}

\item{result}{Pre-computed clustering results used for further splitting. Default NULL.}

\item{min.cells}{Minimal number of cells in a cluster. Default 4}

\item{...}{Other parameters passed to merge_cl}
}
\value{
A list with cluster membership, and top pairwise marker genes.
}
\description{
Iterative consensus clustering
}
