% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_impute.R
\name{iter_impute_knn}
\alias{iter_impute_knn}
\title{Title}
\usage{
iter_impute_knn(knn, ref, dat, tol = 10^-3, max.iter = 100, ...)
}
\arguments{
\item{knn}{}

\item{ref}{}

\item{dat}{}

\item{tol}{}

\item{max.iter}{}

\item{...}{}
}
\description{
Title
}
