% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{jaccard}
\alias{jaccard}
\title{Compute jaccard distances between columns of a matrix}
\usage{
jaccard(m)
}
\arguments{
\item{m}{A matrix or sparse matrix}
}
\value{
a sparse matrix of class dgTMatrix containing Jaccard distances between every pair of samples.
}
\description{
Compute jaccard distances between columns of a matrix
}
