% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{jaccard_leiden}
\alias{jaccard_leiden}
\title{Perform Jaccard/Leiden clustering}
\usage{
jaccard_leiden(
  dat,
  k = 10,
  weight = NULL,
  num_iter = 2,
  resolution_parameter = 1e-04,
  random_seed = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{dat}{A matrix of samples (rows) x features (columns)}

\item{k}{K nearest neighbors to use}

\item{num_iter}{If n_iter < 0 the optimiser continues iterating until it encounters an iteration that did not improve the partition.}
}
\value{
A list object with the cluster factor object and (cl) and Jaccard/Leiden results (result)
}
\description{
Perform Jaccard/Leiden clustering
}
