% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{jaccard_louvain.FNN}
\alias{jaccard_louvain.FNN}
\title{Find nearest neighbors using Jaccard/Louvain metrics}
\usage{
jaccard_louvain.FNN(dat, k = 10, knn.matrix = NULL)
}
\arguments{
\item{knn.matrix}{}
}
\description{
Find nearest neighbors using Jaccard/Louvain metrics
}
