% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{jaccard_louvain.RANN}
\alias{jaccard_louvain.RANN}
\title{Perform Jaccard/Louvain clustering based on RANN}
\usage{
jaccard_louvain.RANN(dat, k = 10)
}
\arguments{
\item{dat}{A matrix of features (rows) x samples (columns)}

\item{k}{K nearest neighbors to use}
}
\value{
A list object with the cluster factor object and (cl) and Jaccard/Louvain results (result)
}
\description{
Perform Jaccard/Louvain clustering based on RANN
}
