% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{jaccard_louvain}
\alias{jaccard_louvain}
\title{Perform Jaccard/Louvain clustering using Rphenograph}
\usage{
jaccard_louvain(dat, k = 10)
}
\arguments{
\item{dat}{A matrix of features (rows) x samples (columns)}

\item{k}{K-nearest neighbors to use for clustering}
}
\value{
A list object with the cluster factor object and (cl) and Rphenograph results (result)
}
\description{
Perform Jaccard/Louvain clustering using Rphenograph
}
