% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{knn_jaccard}
\alias{knn_jaccard}
\title{Compute a jaccard distance matrix based on a matrix of K-Nearest Neigbors}
\usage{
knn_jaccard(knn)
}
\arguments{
\item{knn}{a matrix or sparse matrix of k-nearest neighbors}
}
\value{
a sparse matrix of Jaccard distances produced by the \code{jaccard()} function.
}
\description{
Compute a jaccard distance matrix based on a matrix of K-Nearest Neigbors
}
