% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{knn_joint}
\alias{knn_joint}
\title{knn joint}
\usage{
knn_joint(
  comb.dat,
  ref.sets = names(comb.dat$dat.list),
  select.sets = names(comb.dat$dat.list),
  merge.sets = ref.sets,
  select.cells = comb.dat$all.cells,
  select.genes = NULL,
  method = "cor",
  self.method = "RANN",
  k = 15,
  sample.size = 5000,
  cl.sample.size = 100,
  batch.size = 10000,
  verbose = TRUE,
  mc.cores = 1,
  ...
)
}
\arguments{
\item{comb.dat}{}

\item{ref.sets}{}

\item{select.sets}{}

\item{merge.sets}{}

\item{select.cells}{}

\item{select.genes}{}

\item{method}{}

\item{self.method}{}

\item{k}{}

\item{sample.size}{}

\item{cl.sample.size}{}

\item{batch.size}{}

\item{verbose}{}

\item{mc.cores}{}

\item{...}{}
}
\value{

}
\description{
knn joint
}
