% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{lm_matrix}
\alias{lm_matrix}
\title{Perform a linear model fit on a matrix of data}
\usage{
lm_matrix(dat, x)
}
\arguments{
\item{dat}{a matrix}

\item{x}{a vector}
}
\value{
a list with fit and residuals
}
\description{
Perform a linear model fit on a matrix of data
}
