% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{lm_normalize}
\alias{lm_normalize}
\title{add batch substracted median}
\usage{
lm_normalize(dat, x, R_2.th = 0.2, padj.th = 0.01, min.genes = 5)
}
\arguments{
\item{dat}{}

\item{x}{}

\item{R_2.th}{}

\item{padj.th}{}

\item{min.genes}{}
}
\description{
add batch substracted median
}
