% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{logCPM}
\alias{logCPM}
\title{Convert a matrix of raw counts to a matrix of log2(Counts per Million + 1) values}
\usage{
logCPM(counts)
}
\arguments{
\item{counts}{a matrix, dgCMatrix, or dgTMatrix of count values.}
}
\value{
a matrix, dgCMatrix, or dgTMatrix of log2(CPM + 1) values (matching input)
}
\description{
The input can be a base R matrix or a sparse matrix from the Matrix package.
}
\details{
This function expects that columns correspond to samples, and rows to genes.
}
