% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions_rename.r
\name{makeColorsUnique}
\alias{makeColorsUnique}
\title{Make a vector of unique colors}
\usage{
makeColorsUnique(colorVector, seed = 1)
}
\arguments{
\item{colorVector}{a vector of colors (in a format compatable with col2rgb)}

\item{seed}{random seem for reproducibility}
}
\value{
a color vector very close to the initial colorVector, but
  where all color values are unique
}
\description{
This function converts any vector of colors to a new vector where each
  element is unique but the elements overall are as close as possible to
  the original elements.
}
