% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions_rename.r
\name{makeLayerLabel}
\alias{makeLayerLabel}
\title{Numeric layer for annotation object}
\usage{
makeLayerLabel(Samp.dat, column = "roi_label")
}
\arguments{
\item{Samp.dat}{annotation data frame or tibble in standard format (for human data)}

\item{column}{name of column that contains layer information in format "REGION_L#"}
}
\value{
a numeric vector corresponding to cortical layer (or NA for subcortical regions)
}
\description{
This function expects a specific format for input that may not make sense for most users.
}
