% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions_rename.r
\name{makeRegionLabel}
\alias{makeRegionLabel}
\title{Split region from roi for annotation object}
\usage{
makeRegionLabel(Samp.dat, column = "roi_label")
}
\arguments{
\item{Samp.dat}{annotation data frame or tibble in standard format (for human data)}

\item{column}{name of column that contains layer information in format "REGION_L#"}
}
\value{
a chracter vector corresponding to brain region
}
\description{
This function expects a specific format for input that may not make sense for most users.
}
