% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{map_by_cor}
\alias{map_by_cor}
\title{Map samples to a training dataset by correlation}
\usage{
map_by_cor(train.dat, train.cl, test.dat, method = "median")
}
\arguments{
\item{train.dat}{Training data matrix, usually log-transformed CPM}

\item{train.cl}{Training cluster factor object}

\item{test.dat}{Data for cells to map to the training set. Should have the same genes as train.dat.}

\item{method}{Which statistic to compare. "median" or "mean". Default is "median".}
}
\value{
a list object containing two objects:
\itemize{
\item pred.df: a data.frame with two columns, pred.cl and pred.score with the predicted cluster and correlation scores.
\item cor.matrix: a matrix object with correlation scores for each cluster.
}
}
\description{
Map samples to a training dataset by correlation
}
