% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{map_cl_summary}
\alias{map_cl_summary}
\title{Map a dataset to a reference, and compare existing cluster calls to the reference comparison}
\usage{
map_cl_summary(ref.dat, ref.cl, map.dat, map.cl, method = "median")
}
\arguments{
\item{ref.dat}{Training data matrix, usually log-transformed CPM}

\item{ref.cl}{Training cluster factor object}

\item{map.dat}{Data for cells to map to the training set. Should have the same genes as train.dat.}

\item{map.cl}{Cluster assignments for the training set to compare to results of mapping.}

\item{method}{Which statistic to compare. "median" or "mean". Default is "median".}
}
\value{
a list object with two objects:  
\itemize{
\item map.df: A data.frame with the mapping results for each sample in map.dat to the reference
\item cl.map.df: A data.frame with cluster-level frequency of mapping for each cluster in map.cl to ref.cl
}
}
\description{
Map a dataset to a reference, and compare existing cluster calls to the reference comparison
}
