% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{map_cv}
\alias{map_cv}
\title{Run a single round of cross-validation of cluster mapping using a subset of marker genes}
\usage{
map_cv(
  norm.dat,
  cl,
  markers,
  n.bin = 5,
  g.perc = 1,
  method = "median",
  verbose = TRUE
)
}
\arguments{
\item{norm.dat}{a normalized data matrix for clustered cells}

\item{cl}{a cluster factor object for the cells in norm.dat}

\item{markers}{a character object with the marker genes to use for cross-validation}

\item{n.bin}{an integer indicating the number of bins to use}

\item{g.perc}{the fraction of genes to use for validation.}

\item{method}{Method for mapping. Must be either "median" (Default) or "mean".}

\item{verbose}{Whether or not to display progress notifications.}
}
\value{
a named character object with the results of one round of cross-validation
}
\description{
Run a single round of cross-validation of cluster mapping using a subset of marker genes
}
