% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_dend.R
\name{map_dend}
\alias{map_dend}
\title{Title}
\usage{
map_dend(
  dend,
  cl,
  cl.med,
  dat,
  map.dat,
  select.cells,
  p = 0.8,
  low.th = 0.2,
  default.markers = NULL
)
}
\arguments{
\item{dend}{}

\item{cl}{}

\item{cl.med}{}

\item{dat}{}

\item{map.dat}{}

\item{select.cells}{}

\item{p}{}

\item{low.th}{}

\item{default.markers}{}
}
\description{
Title
}
