% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dend.markers.R
\name{map_dend_markers}
\alias{map_dend_markers}
\title{Map dend markers}
\usage{
map_dend_markers(dend.list, map.dat, select.cells, th = 0.5)
}
\arguments{
\item{dend.list}{}

\item{map.dat}{}

\item{select.cells}{}

\item{th}{}
}
\value{

}
\description{
Map dend markers
}
