% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_dend.R
\name{map_dend_membership}
\alias{map_dend_membership}
\title{Title}
\usage{
map_dend_membership(
  dend,
  cl,
  cl.med,
  dat,
  map.dat,
  map.cells,
  mc.cores = 1,
  bs.num = 100,
  ...
)
}
\arguments{
\item{dend}{}

\item{cl}{}

\item{cl.med}{}

\item{dat}{}

\item{map.dat}{}

\item{map.cells}{}

\item{mc.cores}{}

\item{bs.num}{}

\item{...}{}
}
\description{
Title
}
