% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{map_sampling}
\alias{map_sampling}
\title{Perform bootstrapped mapping using a fraction of provided marker genes.}
\usage{
map_sampling(
  train.dat,
  train.cl,
  test.dat,
  markers,
  markers.perc = 0.8,
  iter = 100,
  method = "median",
  verbose = TRUE
)
}
\arguments{
\item{train.dat}{Training data matrix, usually log-transformed CPM}

\item{train.cl}{Training cluster factor object}

\item{test.dat}{Data for cells to map to the training set. Should have the same genes as train.dat.}

\item{markers}{A vector of marker gene symbols to use for comparisons}

\item{markers.perc}{The fraction of randomly sampled markers to use for each round. Default = 0.8.}

\item{iter}{Number of iterations to perform. Default = 100.}

\item{method}{Method for mapping, passed to map_by_cor(). Default = "median".}

\item{verbose}{Whether or not to display progress notifications.}
}
\value{
a list object with two objects:  
\itemize{
\item map.df: A data.frame with the mapping results for each sample in test.dat to the reference
\item map.freq: A table with the frequency of mapping of each sample to each cluster across all iterations.
}
}
\description{
Perform bootstrapped mapping using a fraction of provided marker genes.
}
