% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_dend.R
\name{mapping}
\alias{mapping}
\title{Title}
\usage{
mapping(
  dend,
  cl.df,
  cl,
  norm.dat,
  query.dat,
  bp.collapse.th = NULL,
  mc.cores = 1,
  bs.num = 100,
  p = 0.7,
  low.th = 0.15,
  conf.th = 0.7,
  min.genes = 1,
  min.genes.ratio = 0.3
)
}
\arguments{
\item{dend}{}

\item{cl.df}{}

\item{cl}{}

\item{norm.dat}{}

\item{query.dat}{}

\item{bp.collapse.th}{}

\item{mc.cores}{}

\item{bs.num}{}

\item{p}{}

\item{low.th}{}

\item{conf.th}{}

\item{min.genes}{}

\item{min.genes.ratio}{}
}
\value{

}
\description{
Title
}
