% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{markers_max_tau}
\alias{markers_max_tau}
\title{Title}
\usage{
markers_max_tau(cl.dat, th = 0.5, tau.th = 0.7, n.markers = 3)
}
\arguments{
\item{cl.dat}{}

\item{th}{}

\item{tau.th}{}

\item{n.markers}{}
}
\description{
Title
}
