% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{markers_tau_one_vs_other}
\alias{markers_tau_one_vs_other}
\title{Title}
\usage{
markers_tau_one_vs_other(
  cl,
  cl.present.counts,
  present.th = 0.4,
  tau.th = 0.8,
  top.n = 10
)
}
\arguments{
\item{cl}{}

\item{cl.present.counts}{}

\item{present.th}{}

\item{tau.th}{}

\item{top.n}{}
}
\description{
Title
}
