% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{match_cl}
\alias{match_cl}
\title{Compute correlations of clusters to a reference set, and get the best-correlated reference clusters}
\usage{
match_cl(cl, dat, ref.cl, ref.cl.df, ref.dat, rename = TRUE)
}
\arguments{
\item{cl}{A cluster factor object to compare to a reference}

\item{dat}{a normalized data matrix for data to map to the reference}

\item{ref.cl}{A cluster factor object for the reference clusters}

\item{ref.cl.df}{A data.frame with reference cluster annotations that includes a "cluster_label" column}

\item{ref.dat}{a reference normalized data matrix}

\item{rename}{a logical indicating whether or not to rename the results.

#' @return a list with 3 objects:  
\itemize{
\item cl
\item cl.df
\item cor
}}
}
\description{
match_cl uses pearson correlation of cluster means.
}
