% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_merge.R
\name{merge_cl_multiple}
\alias{merge_cl_multiple}
\title{Title}
\usage{
merge_cl_multiple(
  comb.dat,
  merge.dat.list,
  cl,
  anchor.genes,
  verbose = TRUE,
  pairBatch = 40,
  de.genes.list = NULL,
  lfc.conservation.th = 0.7,
  merge.type = "undirectional"
)
}
\arguments{
\item{comb.dat}{}

\item{merge.dat.list}{}

\item{cl}{}

\item{anchor.genes}{}

\item{verbose}{}

\item{pairBatch}{}

\item{de.genes.list}{}

\item{lfc.conservation.th}{}

\item{merge.type}{}
}
\description{
Title
}
