% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_plot.R
\name{multiplot}
\alias{multiplot}
\title{Title}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{}

\item{plotlist}{}

\item{file}{}

\item{cols}{}

\item{layout}{}
}
\description{
Title
}
