% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dend.markers.R
\name{node_specific_markers}
\alias{node_specific_markers}
\title{Node specific markers}
\usage{
node_specific_markers(dend.list, norm.dat, cl, ...)
}
\arguments{
\item{dend.list}{}

\item{norm.dat}{}

\item{cl}{}

\item{...}{}
}
\value{

}
\description{
Node specific markers
}
