% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dend.markers.R
\name{node_vs_sibling_markers}
\alias{node_vs_sibling_markers}
\title{Node vs sibling markers}
\usage{
node_vs_sibling_markers(dend.list, norm.dat, cl, ...)
}
\arguments{
\item{dend.list}{}

\item{norm.dat}{}

\item{cl}{}

\item{...}{}
}
\value{

}
\description{
Node vs sibling markers
}
