% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{onestep_clust}
\alias{onestep_clust}
\title{One round of clustering in the iteractive pipeline}
\usage{
onestep_clust(norm.dat, select.cells = colnames(norm.dat), counts = NULL,
  method = c("louvain", "ward"), dim.method = c("pca", "WGCNA"),
  min.cells = 4, vg.padj.th = 0.5, max.dim = 20, rm.gene.mod = NULL,
  rm.eigen = NULL, rm.th = 0.6, de.param = de_param(), min.genes = 5,
  type = "undirectional", maxGenes = 3000, sampleSize = 4000,
  max.cl.size = 300, prefix = NULL, verbose = FALSE)
}
\arguments{
\item{norm.dat}{}

\item{select.cells}{}

\item{counts}{}

\item{method}{}

\item{dim.method}{}

\item{min.cells}{}

\item{vg.padj.th}{}

\item{max.dim}{}

\item{rm.gene.mod}{}

\item{rm.eigen}{}

\item{rm.th}{}

\item{de.param}{}

\item{min.genes}{}

\item{type}{}

\item{maxGenes}{}

\item{sampleSize}{}

\item{max.cl.size}{}

\item{prefix}{}

\item{verbose}{}
}
\description{
One round of clustering in the iteractive pipeline
}
