% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{pair_cor}
\alias{pair_cor}
\title{Compute correlation each row of matrix1 with the corresponding row of matrix2 
matrix1 and matrix2 must have the same dimemsion.}
\usage{
pair_cor(mat1, mat2)
}
\description{
Compute correlation each row of matrix1 with the corresponding row of matrix2 
matrix1 and matrix2 must have the same dimemsion.
}
