% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{pass_louvain}
\alias{pass_louvain}
\title{Test louvain modularity scores based on an adjacency matrix}
\usage{
pass_louvain(mod.sc, adj.mat)
}
\arguments{
\item{mod.sc}{a vector of modularity scores}

\item{adj.mat}{an adjacency matrix}
}
\value{
a logical vector
}
\description{
Test louvain modularity scores based on an adjacency matrix
}
