% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{phenograph}
\alias{phenograph}
\title{Perform PhenoGraph clustering on a matrix of data}
\usage{
phenograph(data, k = 30, verbose = FALSE)
}
\arguments{
\item{data}{a matrix with samples as rows and features as columns.}

\item{k}{an integer, number of neighbors to use. Default = 30.}

\item{verbose}{Whether or not to display status messages. Default is FALSE.}
}
\value{
a list object with two results: g, an igraph graph object; community, an igraph communities object.
}
\description{
This is an R-only reimplementation of the RPhenograph package, which uses C++ for the jaccard coefficient step. 
It also has a verbosity option, which makes it more useful for pipeline analysis.
}
\details{
For the original implementation, see: https://github.com/JinmiaoChenLab/Rphenograph
}
