% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{phenograph_jaccard_coeff}
\alias{phenograph_jaccard_coeff}
\title{Compute Jaccard Coefficients from a K nearest neighbors matrix}
\usage{
phenograph_jaccard_coeff(idx)
}
\arguments{
\item{idx}{a matrix of K-nearest neighbors. Each row contains the K neighbors for a sample.}
}
\value{
a matrix of 3 columns with from, to, and weight for the coefficients
}
\description{
This is an R implementation of the C++ version provided by the Rphenograph package.
}
\details{
See: https://github.com/JinmiaoChenLab/Rphenograph/blob/master/src/jaccard_coeff.cpp
}
