% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_plot.R
\name{plot_3d_label}
\alias{plot_3d_label}
\title{Title}
\usage{
plot_3d_label(
  df,
  col,
  label_col = NULL,
  cex = 1,
  label.cex = cex,
  init = TRUE,
  prefix = NULL,
  bg.col = "gray60"
)
}
\arguments{
\item{df}{}

\item{col}{}

\item{label_col}{}

\item{cex}{}

\item{label.cex}{}

\item{init}{}

\item{prefix}{}

\item{bg.col}{}
}
\description{
Title
}
