% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_plot.R
\name{plot_3d_label_multiple}
\alias{plot_3d_label_multiple}
\title{Title}
\usage{
plot_3d_label_multiple(
  df,
  cols,
  label_cols,
  cex = 0.7,
  label.cex = 0.7,
  fn = NULL,
  win.dim = c(20, 40, 1200, 800),
  layout = NULL,
  bg.col = "gray60",
  dir = "./"
)
}
\arguments{
\item{df}{}

\item{cols}{}

\item{label_cols}{}

\item{cex}{}

\item{label.cex}{}

\item{fn}{}

\item{win.dim}{}

\item{layout}{}

\item{bg.col}{}

\item{dir}{}
}
\description{
Title
}
