% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_plot.R
\name{plot_3d_select}
\alias{plot_3d_select}
\title{Title}
\usage{
plot_3d_select(
  rd.dat,
  select.cells,
  fg.col = "red",
  bg.col = "gray",
  fg.alpha = 1,
  bg.alpha = 0.5,
  cex = 0.15,
  web.fn = NULL,
  web.dir = "./"
)
}
\arguments{
\item{rd.dat}{}

\item{select.cells}{}

\item{fg.col}{}

\item{bg.col}{}

\item{fg.alpha}{}

\item{bg.alpha}{}

\item{cex}{}

\item{web.fn}{}

\item{web.dir}{}
}
\value{

}
\description{
Title
}
