% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_plot.R
\name{plot_3d_val}
\alias{plot_3d_val}
\title{Title}
\usage{
plot_3d_val(
  df,
  val,
  cex = 1,
  max.val = quantile(val, 0.99),
  init = TRUE,
  prefix = NULL,
  bg.col = "gray60"
)
}
\arguments{
\item{df}{}

\item{val}{}

\item{cex}{}

\item{max.val}{}

\item{init}{}

\item{prefix}{}

\item{bg.col}{}
}
\description{
Title
}
