% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap4.R
\name{plot_cl_heatmap4}
\alias{plot_cl_heatmap4}
\title{Title}
\usage{
plot_cl_heatmap4(
  norm.dat,
  cl,
  markers,
  prefix = NULL,
  hc = NULL,
  gene.hc = NULL,
  centered = FALSE,
  labels = names(cl),
  sorted = FALSE,
  by.cl = TRUE,
  ColSideColors = NULL,
  maxValue = 5,
  min.sep = 4,
  main = "",
  height = 13,
  width = 9,
  legend.dat = NULL
)
}
\arguments{
\item{norm.dat}{}

\item{cl}{}

\item{markers}{}

\item{prefix}{}

\item{hc}{}

\item{gene.hc}{}

\item{centered}{}

\item{labels}{}

\item{sorted}{}

\item{by.cl}{}

\item{ColSideColors}{}

\item{maxValue}{}

\item{min.sep}{}

\item{main}{}

\item{height}{}

\item{width}{}

\item{legend.dat}{}
}
\value{

}
\description{
Title
}
