% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_noise_cl.R
\name{plot_cl_low}
\alias{plot_cl_low}
\title{Plot cl low}
\usage{
plot_cl_low(norm.dat, cl, low.df, de.genes, all.col)
}
\arguments{
\item{norm.dat}{}

\item{cl}{}

\item{low.df}{}

\item{de.genes}{}

\item{all.col}{}
}
\value{

}
\description{
Plot cl low
}
