% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_util.R
\name{plot_confusion}
\alias{plot_confusion}
\title{Title}
\usage{
plot_confusion(
  consensus.cl,
  prefix,
  comb.dat,
  consensus.cl.df = NULL,
  do.droplevels = FALSE,
  ...
)
}
\arguments{
\item{consensus.cl}{}

\item{prefix}{}

\item{comb.dat}{}

\item{consensus.cl.df}{}

\item{do.droplevels}{}

\item{...}{}
}
\description{
Title
}
