% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{plot_de_lfc_num}
\alias{plot_de_lfc_num}
\title{Title}
\usage{
plot_de_lfc_num(de.genes, top.n = 100, select.pair = NULL, cl.label = NULL)
}
\arguments{
\item{de.genes}{}

\item{top.n}{}

\item{select.pair}{}

\item{cl.label}{}
}
\description{
Title
}
