% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{plot_de_num}
\alias{plot_de_num}
\title{Title}
\usage{
plot_de_num(
  de.genes,
  dend,
  cl.label = NULL,
  directed = FALSE,
  file = "log10.de.num.pdf",
  ...
)
}
\arguments{
\item{de.genes}{}

\item{dend}{}

\item{cl.label}{}

\item{directed}{}

\item{file}{}

\item{...}{}
}
\value{

}
\description{
Title
}
