% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{plot_dend}
\alias{plot_dend}
\title{Plot dendrogram}
\usage{
plot_dend(dend, dendro_data = NULL, node_size = 1, r = c(-0.1, 1))
}
\arguments{
\item{dend}{}

\item{dendro_data}{}

\item{node_size}{}

\item{r}{}
}
\description{
Plot dendrogram
}
