% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_noise_cl.R
\name{plot_doublet}
\alias{plot_doublet}
\title{Plot doublet}
\usage{
plot_doublet(norm.dat, cl, nn.df, de.genes, all.col)
}
\arguments{
\item{norm.dat}{}

\item{cl}{}

\item{nn.df}{}

\item{de.genes}{}

\item{all.col}{}
}
\value{

}
\description{
Plot doublet
}
