% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{plot_low_qc}
\alias{plot_low_qc}
\title{Plot low-quality clusters}
\usage{
plot_low_qc(norm.dat, cl, low.df, nn.df, de.genes, all.col)
}
\arguments{
\item{cl}{A cluster factor object for all cells.}

\item{low.df}{a data.frame with results from find_low_quality_cl().}

\item{nn.df}{}

\item{de.genes}{differential gene expression results like those generated by de_score().}

\item{all.col}{}

\item{norm.data}{normalized data matrix for clustered cells.}
}
\value{

}
\description{
Plot low-quality clusters
}
