% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_util.R
\name{plot_markers}
\alias{plot_markers}
\title{Title}
\usage{
plot_markers(
  dat.list,
  cl,
  de.param.list,
  prefix,
  common.genes,
  comb.de.genes = NULL,
  cl.means.list = NULL,
  col.list = NULL,
  cl.col = NULL,
  select.genes = NULL,
  save.matrix = FALSE,
  n.markers = 20,
  ...
)
}
\arguments{
\item{dat.list}{}

\item{cl}{}

\item{de.param.list}{}

\item{prefix}{}

\item{common.genes}{}

\item{comb.de.genes}{}

\item{cl.means.list}{}

\item{col.list}{}

\item{cl.col}{}

\item{select.genes}{}

\item{save.matrix}{}

\item{n.markers}{}

\item{...}{}
}
\value{

}
\description{
Title
}
