% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{plot_markers_cl_means}
\alias{plot_markers_cl_means}
\title{Plot markers cl means}
\usage{
plot_markers_cl_means(
  select.genes,
  gene.ordered = FALSE,
  cl.means.list = NULL,
  comb.dat = NULL,
  cl = NULL,
  cl.col = NULL,
  prefix = "",
  ...
)
}
\arguments{
\item{select.genes}{}

\item{gene.ordered}{}

\item{cl.means.list}{}

\item{comb.dat}{}

\item{cl}{}

\item{cl.col}{}

\item{prefix}{}

\item{...}{}
}
\description{
Plot markers cl means
}
