% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{plot_pair_matrix}
\alias{plot_pair_matrix}
\title{Title}
\usage{
plot_pair_matrix(
  pair.num,
  file,
  directed = FALSE,
  dend = NULL,
  col = jet.colors(100),
  cl.label = NULL,
  ...
)
}
\arguments{
\item{pair.num}{}

\item{file}{}

\item{directed}{}

\item{dend}{}

\item{col}{}

\item{cl.label}{}

\item{...}{}
}
\description{
Title
}
