% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_plot.R
\name{plot_RD_cl}
\alias{plot_RD_cl}
\title{Title}
\usage{
plot_RD_cl(
  rd.dat,
  cl,
  cl.color,
  cl.label,
  cex = 0.15,
  fn.size = 2,
  alpha.val = 1,
  show.legend = FALSE,
  legend.size = 2,
  label.center = TRUE,
  bg = "blank",
  fn.color = "black"
)
}
\arguments{
\item{rd.dat}{}

\item{cl}{}

\item{cl.color}{}

\item{cl.label}{}

\item{cex}{}

\item{fn.size}{}

\item{alpha.val}{}

\item{show.legend}{}

\item{legend.size}{}

\item{label.center}{}

\item{bg}{}

\item{fn.color}{}
}
\value{

}
\description{
Title
}
