% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_plot.R
\name{plot_RD_meta}
\alias{plot_RD_meta}
\title{Title}
\usage{
plot_RD_meta(
  rd.dat,
  meta,
  meta.col = NULL,
  show.legend = TRUE,
  cex = 0.15,
  legend.size = 5,
  alpha.val = 1
)
}
\arguments{
\item{rd.dat}{}

\item{meta}{}

\item{meta.col}{}

\item{show.legend}{}

\item{cex}{}

\item{legend.size}{}

\item{alpha.val}{}
}
\description{
Title
}
