% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R, R/harmonize_util.R
\name{plot_tsne}
\alias{plot_tsne}
\title{Plot tsne}
\usage{
plot_tsne(
  cl,
  cl.df,
  comb.dat,
  prefix,
  tsne.df,
  cex = 0.3,
  fn.size = 2,
  height = 8,
  width = 10
)

plot_tsne(
  cl,
  cl.df,
  comb.dat,
  prefix,
  tsne.df,
  cex = 0.3,
  fn.size = 2,
  height = 8,
  width = 10
)
}
\arguments{
\item{cl}{}

\item{cl.df}{}

\item{comb.dat}{}

\item{prefix}{}

\item{tsne.df}{}

\item{cex}{}

\item{fn.size}{}

\item{height}{}

\item{width}{}
}
\description{
Plot tsne

Title
}
