% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_tSNE.R
\name{plot_tsne_cl}
\alias{plot_tsne_cl}
\title{Title}
\usage{
plot_tsne_cl(
  norm.dat,
  select.genes,
  cl,
  cl.df,
  tsne.df = NULL,
  show.legend = FALSE,
  cex = 0.15,
  fn.size = 2,
  alpha.val = 1,
  ...
)
}
\arguments{
\item{norm.dat}{}

\item{select.genes}{}

\item{cl}{}

\item{cl.df}{}

\item{tsne.df}{}

\item{show.legend}{}

\item{cex}{}

\item{fn.size}{}

\item{alpha.val}{}

\item{...}{}
}
\value{

}
\description{
Title
}
