% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vg.R
\name{plot_vg}
\alias{plot_vg}
\title{Generate gene variance plots}
\usage{
plot_vg(gene_var_stats, plots = "all", loess_fit = NULL)
}
\arguments{
\item{gene_var_stats}{a data.frame of statistics generated by compute_vg_stats()}

\item{plots}{a character vector specifying one or more of the following options (Default is "all"):
\itemize{
  \item qq_z_plot
  \item qq_loess.z_plot
  \item z_density_plot
  \item loess.z_density_plot
  \item dispersion_fit_plot
  \item all
}}

\item{loess_fit}{a loess fit object as generated by gene_loess_fit(). This is required only for the dispersion_fit_plot.}
}
\value{
a list object with multiple ggplot2 plot objects depending on the plots parameter
}
\description{
Generate gene variance plots
}
