% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_PCA.R
\name{prcomp.irlba}
\alias{prcomp.irlba}
\title{prcomp irlba}
\usage{
\method{prcomp}{irlba}(x, max.rank = 500, maxit = 1000, tol = 1e-05, center = TRUE, ...)
}
\arguments{
\item{x}{}

\item{max.rank}{}

\item{maxit}{}

\item{tol}{}

\item{center}{}

\item{...}{}
}
\value{

}
\description{
prcomp irlba
}
