% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{predict_annotate_cor}
\alias{predict_annotate_cor}
\title{Predict annotations by cluster correlation}
\usage{
predict_annotate_cor(
  cl,
  norm.dat,
  ref.markers,
  ref.cl,
  ref.cl.df,
  ref.norm.dat,
  method = "median",
  reorder = TRUE
)
}
\arguments{
\item{cl}{a cluster factor object for data to map to the reference}

\item{norm.dat}{a normalized data matrix for data to map to the reference}

\item{ref.markers}{a set of reference marker genes}

\item{ref.cl}{a reference cluster factor object}

\item{ref.cl.df}{a reference cl.df data.frame that describes the reference clusters}

\item{ref.norm.dat}{a reference normalized data matrix}

\item{method}{"median" or "mean". Default is "median".}

\item{reorder}{Whether or not to reorder the input clusters based on the reference.}
}
\value{
a list object with annotation results
}
\description{
This function performs map_by_cor(), then compare_annotate().
}
