% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_impute.R
\name{predict_knn_new}
\alias{predict_knn_new}
\title{Title}
\usage{
predict_knn_new(knn.idx, reference, cl, ...)
}
\arguments{
\item{knn.idx}{}

\item{reference}{}

\item{cl}{}

\item{...}{}
}
\description{
Title
}
