% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{prune_dend}
\alias{prune_dend}
\title{Prune dendrogram}
\usage{
prune_dend(dend, rm.labels, top.level = TRUE)
}
\arguments{
\item{dend}{}

\item{rm.labels}{}

\item{top.level}{}
}
\description{
Prune dendrogram
}
