% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{pvclust_show_signif_gradient}
\alias{pvclust_show_signif_gradient}
\title{Pv clust show significant gradient}
\usage{
pvclust_show_signif_gradient(
  dend,
  pvclust_obj,
  signif_type = c("bp", "au"),
  signif_col_fun = colorRampPalette(c("black", "darkred", "red")),
  ...
)
}
\arguments{
\item{dend}{}

\item{pvclust_obj}{}

\item{signif_type}{}

\item{signif_col_fun}{}

\item{...}{}
}
\value{

}
\description{
Pv clust show significant gradient
}
