% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_PCA.R
\name{rd_PCA}
\alias{rd_PCA}
\title{Reduce dimensionality using PCA}
\usage{
rd_PCA(norm.dat, select.genes = NULL, select.cells = NULL,
  sampled.cells = NULL, max.pca = 10, th = 2, w = NULL)
}
\arguments{
\item{norm.dat}{a normalized data matrix.}

\item{select.genes}{a character vector of genes to use for dimensionality reduction.}

\item{select.cells}{a character vector of samples to use for dimensionality reduction.}

\item{sampled.cells}{}

\item{max.pca}{}

\item{th}{}

\item{w}{}
}
\description{
Reduce dimensionality using PCA
}
