% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_WGCNA.R
\name{rd_WGCNA}
\alias{rd_WGCNA}
\title{Compute reduced dimensions using WGCNA}
\usage{
rd_WGCNA(norm.dat, select.genes, select.cells,
  sampled.cells = select.cells, min.module.size = 10,
  cutHeight = 0.99, type = "unsigned", soft.power = 4,
  rm.gene.mod = NULL, rm.eigen = NULL, ...)
}
\arguments{
\item{type}{Character, type parameter passed to \code{WGCNA::adjacency()}. Must be one of "unsigned", "signed", "signed hybrid", or "distance". Default is "unsigned".}

\item{soft.power}{Numeric, power parameter passed to \code{WGCNA::adjacency()}}

\item{...}{Additional parameters passed to \code{filter_gene_mod()}}
}
\value{
a list object with reduced dimensions (rd.dat) and gene modules (gm)
}
\description{
Compute reduced dimensions using WGCNA
}
