% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusCluster.R
\name{refine_cl}
\alias{refine_cl}
\title{Refine clusters}
\usage{
refine_cl(cl, co.ratio = NULL, cl.mat = NULL, confusion.th = 0.6,
  min.cells = 4, niter = 50, tol.th = 0.02, verbose = 0)
}
\arguments{
\item{cl}{Cluster membership cluster}

\item{co.ratio}{cell-cell co-clustering matrix.}

\item{cl.mat}{cell-cluster matrix collected from bootstrapping iterations. Either co.ratio or cl.mat should not be NULL.}

\item{confusion.th}{Clusters with average confusion score greater than this threshold will be removed. Cells in this cluster will be re-distributed to other most likely clusters.}

\item{min.cells}{Clusters with fewer than this many cells will be removed. Cells in this cluster will be re-distributed to other most likely clusters.}

\item{niter}{maxmimal mumber of refinement iterations.}

\item{tol.th}{If improvement is smaller than this threshold, terminate refinement step.}

\item{verbose}{If true, print out step-by-step improvement.}
}
\description{
Refine clusters
}
\author{
Zizhen Yao
}
