% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{reorder_cl}
\alias{reorder_cl}
\title{Reorder cluster based on hiearchical clustering of clusters based on average cluster values for the input data matrix}
\usage{
reorder_cl(cl, dat)
}
\arguments{
\item{cl}{A vector of cluster membership with cells as names, and cluster id as values.}

\item{dat}{The data matrix with cells as columns.}
}
\value{
Reorder cluster membership vector. The cluster id start from 1 to the number of clusters.
}
\description{
Reorder cluster based on hiearchical clustering of clusters based on average cluster values for the input data matrix
}
