% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{reorder_dend}
\alias{reorder_dend}
\title{Reorder dendrogram}
\usage{
reorder_dend(dend, l.rank, top.level = TRUE)
}
\arguments{
\item{dend}{}

\item{l.rank}{}

\item{top.level}{}
}
\description{
Reorder dendrogram
}
