% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions.R
\name{reorder_factor}
\alias{reorder_factor}
\title{Reorder factors of one annotation to match another}
\usage{
reorder_factor(clMatch, clRef)
}
\arguments{
\item{clMatch}{A cluster factor object to compare to a reference}

\item{clRef}{A cluster factor object for the reference clusters}
}
\value{
The `clMatch` vector with levels rearranged to match `clRef` in a 
  reasonable way.
}
\description{
This function takes two identical annotations (e.g, two cluster calls for the 
  same cells) and updates the levels of the match cluster to correspond to 
  the levels of the reference cluster.  This is useful to run prior to
  `compare_plot` if clusters are not already ordered.
  
`clRef` and `clMatch` should be factors, but can be coerced from characters
  if needed.
}
