% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vg.R
\name{rescale_samples}
\alias{rescale_samples}
\title{Rescale a gene x sample matrix}
\usage{
rescale_samples(dat)
}
\arguments{
\item{dat}{a matrix or dg[CT]Matrix of expression values with genes as rows and samples as columns.}
}
\value{
a rescaled matrix of expression values
}
\description{
This function scales each sample/column based on the median of the total of gene expression values for each sample.
}
