% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_dend.R
\name{resolve_cl}
\alias{resolve_cl}
\title{Title}
\usage{
resolve_cl(
  cl.g,
  cl.med,
  markers,
  dat,
  map.dat,
  select.cells,
  p = 0.7,
  low.th = 0.2
)
}
\arguments{
\item{cl.g}{}

\item{cl.med}{}

\item{markers}{}

\item{dat}{}

\item{map.dat}{}

\item{select.cells}{}

\item{p}{}

\item{low.th}{}
}
\description{
Title
}
