% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_consensus.R
\name{run_consensus_clust}
\alias{run_consensus_clust}
\title{Run bootstrapped, iterative consensus clustering}
\usage{
run_consensus_clust(norm.dat, select.cells = colnames(norm.dat),
  niter = 100, sample.frac = 0.8, de.param = de_param(),
  output_dir = "subsample_result", overwrite = FALSE, mc.cores = 1,
  init.result = NULL, ...)
}
\arguments{
\item{norm.dat}{A matrix of normalized data}

\item{select.cells}{A selected subset of sample IDs. Default is colnames(norm.dat).}

\item{niter}{Number of bootstrapping rounds to run. Default is 100.}

\item{sample.frac}{Fraction of samples to use per bootstrapping round. Default is 0.8}

\item{de.param}{DE parameters set using `de_param()`.}

\item{output_dir}{Output directory for intermediate clustering results. Set to NULL to run in memory-only mode. This may take a significant quantity of RAM.}

\item{overwrite}{Overwrite flag for file output. Default is FALSE.}

\item{mc.cores}{Number of cores to use for parallel processing. If set to 1 (default), will use single-core mode. If "auto", will use detected cores - 1.}

\item{init.result}{Initial results.}

\item{...}{Additional parameters passed to `iter_clust()`}
}
\description{
Run bootstrapped, iterative consensus clustering
}
