% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_Umap.R
\name{rd_Umap}
\alias{rd_Umap}
\title{Adapt from Seurat package}
\usage{
\method{RunUMAP}{default}(
  mat,
  n.neighbors = 30L,
  n.components = 2L,
  metric = "correlation",
  n.epochs = NULL,
  learning.rate = 1,
  min.dist = 0.3,
  spread = 1,
  set.op.mix.ratio = 1,
  local.connectivity = 1L,
  repulsion.strength = 1,
  negative.sample.rate = 5,
  a = NULL,
  b = NULL,
  seed.use = 42,
  metric.kwds = NULL,
  angular.rp.forest = FALSE,
  reduction.key = "UMAP_",
  verbose = TRUE
)
}
\description{
Adapt from Seurat package
}
