% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sample_cells_by_genecounts}
\alias{sample_cells_by_genecounts}
\title{Sample cells from each cluster weighted by gene count}
\usage{
sample_cells_by_genecounts(cl, norm.dat, max.cl.size = 200)
}
\arguments{
\item{cl}{A cluster factor object}

\item{norm.dat}{normalized gene expression matrix}

\item{max.cl.size}{The maximum number of cells to sample for each cluster}
}
\value{
A downsampled cluster factor object
}
\description{
Sample cells from each cluster weighted by gene count
}
