% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{sample_sets_list}
\alias{sample_sets_list}
\title{Sample sets lists}
\usage{
sample_sets_list(cells.list, cl.list, cl.sample.size = 100, sample.size = 5000)
}
\arguments{
\item{cells.list}{}

\item{cl.list}{}

\item{cl.sample.size}{}

\item{sample.size}{}
}
\value{

}
\description{
Sample sets lists
}
