% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_WGCNA.R
\name{score_gene_mod}
\alias{score_gene_mod}
\title{Compute scores for a set of gene modules}
\usage{
score_gene_mod(norm.dat, select.cells, gene.mod, eigen = NULL,
  method = "average", de.param = de_param(), max.cl.size = NULL)
}
\arguments{
\item{max.cl.size}{}
}
\description{
Compute scores for a set of gene modules
}
