% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{score_pair_chisq}
\alias{score_pair_chisq}
\title{Perform pairwise differential detection tests using Chi-Squared for a single pair of clusters}
\usage{
score_pair_chisq(pair, cl.present, cl.means, cl.size, genes)
}
\arguments{
\item{pair}{a numeric vector of length 2 specifying which clusters to compare}

\item{cl.present}{a data.frame of gene detection proportions (genes x clusters)}

\item{cl.means}{a data.frame of normalized mean gene expression values (genes x clusters)}

\item{cl.size}{a named numeric vector of cluster sizes}

\item{genes}{the genes to use for pairwise comparisons}
}
\value{
a data.frame with DE statistics:
\itemize{
\item{padj} P-values adjusted using the Holm (1979) method (\code{p.adjust()} default).
\item{pval} P-values reported by the \code{vec_chisq_test()} function.
\item{lfc} Log fold change of mean expression values between the pair of clusters.
\item{meanA} Mean expression value for the first cluster in the pair.
\item{meanB} Mean expression value for the second cluster in the pair.
\item{q1} Proportion of cells expressing each gene for the first cluster in the pair.
\item{q2} Proportion of cells expressing each gene for the second cluster in the pair.
}
}
\description{
Perform pairwise differential detection tests using Chi-Squared for a single pair of clusters
}
