% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{score_pair_limma}
\alias{score_pair_limma}
\title{Perform pairwise DE tests using limma for a single pair of clusters}
\usage{
score_pair_limma(pair, cl.present, cl.means, design, fit, genes)
}
\arguments{
\item{pair}{a numeric vector of length 2 specifying which clusters to compare}

\item{cl.present}{a data.frame of gene detection proportions (genes x clusters)}

\item{cl.means}{a data.frame of mean gene expression values (genes x clusters)}

\item{design}{a limma design object}

\item{fit}{a limma fit object}

\item{genes}{the genes to use for pairwise comparisons}
}
\value{
a data.frame with DE statistics:
\itemize{
\item{padj}
\item{pval}
\item{lfc}
\item{meanA}
\item{meanB}
\item{q1}
\item{q2}
}
}
\description{
Perform pairwise DE tests using limma for a single pair of clusters
}
