% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dend.markers.R
\name{select_dend_markers}
\alias{select_dend_markers}
\title{Select dend markers}
\usage{
select_dend_markers(
  dend,
  cl,
  de.genes,
  norm.dat = NULL,
  up.gene.score = NULL,
  down.gene.score = NULL,
  ...
)
}
\arguments{
\item{dend}{}

\item{cl}{}

\item{de.genes}{}

\item{norm.dat}{}

\item{up.gene.score}{}

\item{down.gene.score}{}

\item{...}{}
}
\value{

}
\description{
Select dend markers
}
