% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{select_joint_genes}
\alias{select_joint_genes}
\title{Select joint genes}
\usage{
select_joint_genes(
  comb.dat,
  ref.list,
  select.cells = comb.dat$all.cells,
  maxGenes = 2000,
  vg.padj.th = 0.5,
  max.dim = 20,
  use.markers = TRUE,
  top.n = 100,
  rm.eigen = NULL,
  rm.th = rep(0.7, ncol(rm.eigen))
)
}
\arguments{
\item{comb.dat}{}

\item{ref.list}{}

\item{select.cells}{}

\item{maxGenes}{}

\item{vg.padj.th}{}

\item{max.dim}{}

\item{use.markers}{}

\item{top.n}{}

\item{rm.eigen}{}

\item{rm.th}{}
}
\description{
Select joint genes
}
