% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_util.R
\name{select_joint_markers}
\alias{select_joint_markers}
\title{Title}
\usage{
select_joint_markers(comb.dat, cl, de.genes.list, ...)
}
\arguments{
\item{comb.dat}{}

\item{cl}{}

\item{de.genes.list}{}

\item{...}{}
}
\value{

}
\description{
Title
}
