% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{select_markers}
\alias{select_markers}
\title{Select a set of marker genes based on differential gene expression results.}
\usage{
select_markers(norm.dat, cl, n.markers = 20, de.genes = NULL, ...)
}
\arguments{
\item{norm.dat}{a normalized data matrix for data.}

\item{cl}{a cluster factor object.}

\item{n.markers}{Maximum of markers to return for each cluster (Default = 20)}

\item{de.genes}{Optional, pre-computed DE Genes using de_stats_all_pairs()}

\item{...}{Additional parameters passed to de_stats_all_pairs()}
}
\value{
a list with two objects: markers, a character vector of marker genes; 
de.genes, a list object with differential expression results.
}
\description{
Select a set of marker genes based on differential gene expression results.
}
