% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{select_markers_pair}
\alias{select_markers_pair}
\title{Title}
\usage{
select_markers_pair(
  de.genes,
  add.genes,
  gene.score = NULL,
  rm.genes = NULL,
  top.n = 50,
  max.num = 2000
)
}
\arguments{
\item{de.genes}{}

\item{add.genes}{}

\item{gene.score}{}

\item{rm.genes}{}

\item{top.n}{}

\item{max.num}{}
}
\description{
Title
}
