% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{select_markers_pair_direction}
\alias{select_markers_pair_direction}
\title{Title}
\usage{
select_markers_pair_direction(
  de.genes,
  add.up,
  add.down,
  up.gene.score = NULL,
  down.gene.score = NULL,
  rm.genes = NULL,
  top.n = 50,
  max.num = 2000
)
}
\arguments{
\item{de.genes}{}

\item{add.up}{}

\item{add.down}{}

\item{up.gene.score}{}

\item{down.gene.score}{}

\item{rm.genes}{}

\item{top.n}{}

\item{max.num}{}
}
\description{
Title
}
