% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{select_N_markers}
\alias{select_N_markers}
\title{Title}
\usage{
select_N_markers(
  de.genes,
  up.gene.score = NULL,
  down.gene.score = NULL,
  default.markers = NULL,
  pair.num = 1,
  add.up = pair.num,
  add.down = pair.num,
  rm.genes = NULL,
  pairs = names(de.genes)
)
}
\arguments{
\item{de.genes}{}

\item{up.gene.score}{}

\item{down.gene.score}{}

\item{default.markers}{}

\item{pair.num}{}

\item{add.up}{}

\item{add.down}{}

\item{rm.genes}{}

\item{pairs}{}
}
\description{
Title
}
