% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dend.markers.R
\name{select_node_specific_markers}
\alias{select_node_specific_markers}
\title{Select node specific markers}
\usage{
select_node_specific_markers(
  dend,
  norm.dat,
  cl,
  n.markers = 10,
  de.genes = NULL,
  up.gene.score = NULL,
  down.gene.score = NULL,
  top.n = 50,
  max.num = 2000
)
}
\arguments{
\item{dend}{}

\item{norm.dat}{}

\item{cl}{}

\item{n.markers}{}

\item{de.genes}{}

\item{up.gene.score}{}

\item{down.gene.score}{}

\item{top.n}{}

\item{max.num}{}
}
\value{

}
\description{
Select node specific markers
}
