% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dend.markers.R
\name{select_pos_dend_markers}
\alias{select_pos_dend_markers}
\title{Select pos dend markers}
\usage{
select_pos_dend_markers(dend, cl, norm.dat)
}
\arguments{
\item{dend}{}

\item{cl}{}

\item{norm.dat}{}
}
\value{

}
\description{
Select pos dend markers
}
