% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{select_pos_markers}
\alias{select_pos_markers}
\title{Title}
\usage{
select_pos_markers(
  de.genes,
  cl,
  n.markers = 3,
  default.markers = NULL,
  rm.genes = NULL,
  up.gene.score = NULL,
  down.gene.score = NULL
)
}
\arguments{
\item{de.genes}{}

\item{cl}{}

\item{n.markers}{}

\item{default.markers}{}

\item{rm.genes}{}

\item{up.gene.score}{}

\item{down.gene.score}{}
}
\description{
Title
}
