% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{set_pair_matrix}
\alias{set_pair_matrix}
\title{Update a matrix with values from a 1d vector using row and column positions}
\usage{
set_pair_matrix(m, rows, cols, vals)
}
\arguments{
\item{m}{a matrix object}

\item{rows}{row positions, either as indices or matches to row names}

\item{cols}{row positions, either as indices or matches to row names}

\item{vals}{values to insert at [rows/cols]}
}
\value{
a matrix with updated values
}
\description{
Update a matrix with values from a 1d vector using row and column positions
}
