% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{sparse_cor}
\alias{sparse_cor}
\title{Compute correlation scores for columns of a sparse matrix}
\usage{
sparse_cor(m)
}
\arguments{
\item{m}{a sparse matrix, preferrably dgCMatrix}
}
\value{
a matrix of correlation values between each column of m
}
\description{
Compute correlation scores for columns of a sparse matrix
}
