% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_dend.R
\name{summarize_cl}
\alias{summarize_cl}
\title{Title}
\usage{
summarize_cl(
  dend,
  memb,
  map.dat,
  exp.th = 1,
  conf.th = 0.7,
  min.genes.ratio = 0.3,
  min.genes = 3
)
}
\arguments{
\item{dend}{}

\item{memb}{}

\item{map.dat}{}

\item{exp.th}{}

\item{conf.th}{}

\item{min.genes.ratio}{}

\item{min.genes}{}
}
\value{

}
\description{
Title
}
