% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_cl.R
\name{test_merge}
\alias{test_merge}
\title{Title}
\usage{
test_merge(de.pair, de.param, merge.type = "undirectional")
}
\arguments{
\item{de.pair}{}

\item{de.param}{}

\item{merge.type}{}
}
\description{
Title
}
