% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{unbranch_by_length}
\alias{unbranch_by_length}
\title{Unbranch by length}
\usage{
unbranch_by_length(dend, length.th)
}
\arguments{
\item{dend}{}

\item{length.th}{}
}
\value{

}
\description{
Unbranch by length
}
