% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions_rename.r
\name{updateSampDat}
\alias{updateSampDat}
\title{Update cluster names in annotation data frame}
\usage{
updateSampDat(Samp.dat, clusterInfo)
}
\arguments{
\item{Samp.dat}{Sample information with rows as samples and columns for annotations.}

\item{clusterInfo}{Output cluster information data frame from \code{renameAndOrderClusters}}
}
\value{
Samp.dat variable but with updated cluster names
}
\description{
Update cluster names in annotation data frame
}
