% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{vec_chisq_test}
\alias{vec_chisq_test}
\title{Vectorized Chi-squared tests for differential gene detection}
\usage{
vec_chisq_test(x, x.total, y, y.total)
}
\arguments{
\item{x}{an integer vector with the number of cells in group \emph{x} with detection of each gene.}

\item{x.total}{an integer value with the total number of cells in group \emph{x}.}

\item{y}{an integer vector with the number of cells in group \emph{y} with detection of each gene.}

\item{y.total}{an integer value with the total number of cells in group \emph{y}.}
}
\value{
a data.frame with the following result for each gene:
\itemize{
\item{stats: The value of the chi-squared test statistic}
\item{pval: The p-value as reported by pchisq}
\item{logFC: The log2(fold change) in detection frequency between samples (x / y)}
\item{diff: The difference in proportions between the samples (x - y)}
}
}
\description{
This function uses vectors of the number of samples in two sets that have detection of 
a set of genes and the total number of cells in each set to compute Chi-quared tests with 1 DOF for 
differential detection.
}
