% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions.R
\name{check_donor}
\alias{check_donor}
\title{Find donor clusters}
\usage{
check_donor(
  anno,
  cluster,
  norm.dat,
  select.cells = names(cluster),
  keep.cl = NULL,
  meta1_area = "roi",
  meta2_donor = "external_donor_name",
  plot = TRUE,
  plot.path = "output/"
)
}
\arguments{
\item{anno}{anno dataframe which must include column names listed in `neun.colname`
`meta1_area` and `meta2_donor` below. "cluster" is added from `cluster` parameter below.}

\item{cluster}{clusters labels for all cells along with sample_id as their names}

\item{norm.dat}{expression dataframe with columns as cells and rows as gene names 
and cpm normalized}

\item{select.cells}{column nmaes of norm.dat}

\item{keep.cl}{clusters to definitely keep in analysis (e.g., to exclude from 
consideration as a donor cluster) default is NULL}

\item{meta1_area}{The metadata column that contains information about the area 
(e.g. cortical region and layer).  default is "roi"}

\item{meta2_donor}{The metadata column that contains information about the donor. 
default is "external_donor_name"}

\item{plot}{default is TRUE}

\item{plot.path}{path of plot, default is "./output/"}
}
\value{
Gives cluster ids vs donor expression/region heatmap
}
\description{
This function identifies clusters that are nearly exclusively expressed in one donor.
  Currently it is hard-coded to define clusters with >90% enrichment in a single donor
  or >50% more than expected based on the total number of cells per donor.  Later
  function iterations will include these options as parameters.
}
