% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions.R
\name{check_neun}
\alias{check_neun}
\title{Check whether clusters are neun/non-neun}
\usage{
check_neun(anno, cluster, neun.thresh = 0.5)
}
\arguments{
\item{anno}{anno dataframe with the following required colnames - "facs_population_plan", "cluster"}

\item{cluster}{"cluster" column from anno dataframe}

\item{neun.thresh}{default is 0.5}
}
\value{
returns all the clusters and true/false of whether they are neun positive
}
\description{
Check whether clusters are neun/non-neun
}
\examples{
check_neun(anno, anno$cluster, neun.thresh = 0.5)
}
