% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions.R
\name{check_neun}
\alias{check_neun}
\title{Check whether clusters are Neun+/Neun-}
\usage{
check_neun(anno, cluster, neun.thresh = 0.5,
  neun.colname = "facs_population_plan")
}
\arguments{
\item{anno}{anno dataframe which must include column name listed in `neun.colname`}

\item{cluster}{cluster labels for all cells along with sample_id as their names}

\item{neun.thresh}{fraction of cells expressing NeuN to be considered 
NeuN positive (default is 0.5)}

\item{neun.colname}{column name in `anno` with the Neun information}
}
\value{
returns all the clusters and true/false of whether they are Neun positive
}
\description{
This function is very specific to brain cell/nuclear data where Neun is used as a
  stain for neuronal vs. non-neuronal identity.  This function determine's whether 
  a cluster is predicted to be neuronal or non-neuronal based on expression of Neun.
}
\examples{
check_neun(anno, anno$cluster, neun.thresh = 0.5)
}
