% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions.R
\name{check_outlier}
\alias{check_outlier}
\title{check for outlier clusters}
\usage{
check_outlier(anno, cluster, norm.dat, select.cells = colnames(norm.dat),
  keep.cl = NULL, neun.thresh = 0.5,
  neun.colname = "facs_population_plan",
  qc.metrics = c("Genes.Detected.CPM", "percent_reads_aligned_total",
  "complexity_cg"), test.genes = c("SNAP25", "GAD1", "GAD2", "SLC17A7",
  "SLC17A6"), plot = TRUE, plot.path = "output/")
}
\arguments{
\item{anno}{anno dataframe which must include column names listed in `neun.colname`
and `qc.metrics`.  "cluster" is added from `cluster` parameter below.}

\item{cluster}{cluster labels for all cells along with sample_id as their names}

\item{norm.dat}{expression dataframe with columns as cells and rows as gene names 
and cpm normalized}

\item{select.cells}{column nmaes of norm.dat}

\item{keep.cl}{clusters to definitely keep in analysis (e.g., to exclude from 
consideration as an outlier cluster) default is NULL}

\item{neun.thresh}{fraction of cells expressing NeuN to be considered 
NeuN positive (default is 0.5)}

\item{neun.colname}{column name in anno with the Nuen information}

\item{qc.metrics}{required columns from anno dataframe. default is Genes.Detected.CPM", 
"percent_reads_aligned_total", "complexity_cg"}

\item{test.genes}{CURRENTLY NOT USED.  This function will eventually allow for a 
pre-defined set of genes to be entered. default is "SNAP25", "GAD1", "GAD2", 
"SLC17A7", "SLC17A6"}

\item{plot}{default is TRUE}

\item{plot.path}{path of plot, default is ./output/}
}
\value{
gives outlier clusters and exploratory plots
}
\description{
check for outlier clusters
}
\examples{
check_outlier()
}
