% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions.R
\name{check_outlier}
\alias{check_outlier}
\title{check for outlier clusters}
\usage{
check_outlier(anno, cluster, norm.dat, select.cells = colnames(norm.dat),
  keep.cl = NULL, neun.thresh = 0.5,
  qc.metrics = c("Genes.Detected.CPM", "percent_reads_aligned_total",
  "complexity_cg"), test.genes = c("SNAP25", "GAD1", "GAD2", "SLC17A7",
  "SLC17A6"), plot = TRUE, plot.path = "output/")
}
\arguments{
\item{anno}{anno dataframe with the following required columns - facs_population_plan", "cluster"}

\item{cluster}{clusters labels for all cells along with sample_id as their names}

\item{norm.dat}{expression dataframe with columns as cells and rows as gene names and cpm normalized}

\item{select.cells}{column nmaes of norm.dat}

\item{keep.cl}{default NULL}

\item{neun.thresh}{default 0.5}

\item{qc.metrics}{required columns from anno dataframe. default is Genes.Detected.CPM", "percent_reads_aligned_total", "complexity_cg"}

\item{test.genes}{default is "SNAP25", "GAD1", "GAD2", "SLC17A7", "SLC17A6"}

\item{plot}{default is TRUE}

\item{plot.path}{path of plot, default is ./output/}
}
\value{
gives outlier clusters and exploratory plots
}
\description{
check for outlier clusters
}
\examples{
check_outlier()
}
