% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusCluster.R
\name{collect_co_matrix_sparseM}
\alias{collect_co_matrix_sparseM}
\title{Title}
\usage{
collect_co_matrix_sparseM(norm.dat, result.files, all.cells,
  max.cl.size = 1000)
}
\arguments{
\item{norm.dat}{}

\item{result.files}{}

\item{all.cells}{}

\item{max.cl.size}{}
}
\description{
Title
}
