% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cpm}
\alias{cpm}
\title{Convert a matrix of count values to CPM (Counts per Million)}
\usage{
cpm(counts)
}
\arguments{
\item{counts}{a standard or sparse matrix}
}
\value{
a matrix object of the same type as counts with normalized values
}
\description{
Convert a matrix of count values to CPM (Counts per Million)
}
