% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cpm}
\alias{cpm}
\title{Convert a matrix of raw counts to a matrix of Counts per Million values}
\usage{
cpm(counts)
}
\arguments{
\item{counts}{a matrix of count values.}
}
\description{
The input can be a base R matrix or a sparse matrix from the Matrix package.
}
\details{
This function expects that columns correspond to samples, and rows to genes.
}
