% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{DE_genes_pairs}
\alias{DE_genes_pairs}
\title{Perform pairwise differential gene expression tests between clusters}
\usage{
DE_genes_pairs(norm.dat, cl, pairs, method = "limma", low.th = 1,
  min.cells = 4, cl.present = NULL, use.voom = FALSE,
  counts = NULL, mc.cores = 1)
}
\arguments{
\item{norm.dat}{}

\item{cl}{}

\item{pairs}{}

\item{method}{}

\item{low.th}{}

\item{min.cells}{}

\item{cl.present}{}

\item{use.voom}{}

\item{counts}{}

\item{mc.cores}{}
}
\description{
Perform pairwise differential gene expression tests between clusters
}
