% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{DE_genes_pw}
\alias{DE_genes_pw}
\title{Perform all pairwise differential expression comparison between clusters}
\usage{
DE_genes_pw(norm.dat, cl, ...)
}
\arguments{
\item{...}{Additional parameters passed to DE_genes_pairs()}
}
\value{
a list containing DE results for every pair of clusters
}
\description{
Perform all pairwise differential expression comparison between clusters
}
\seealso{
DE_genes_pairs
}
