% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{de_pair}
\alias{de_pair}
\title{Filter a DE gene results data.frame for significant results}
\usage{
de_pair(df, de.param = de_param(), cl.size1 = NULL, cl.size2 = NULL)
}
\arguments{
\item{cl.size2}{}
}
\description{
Filter a DE gene results data.frame for significant results
}
