% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_function_names.R
\name{de_pair}
\alias{de_pair}
\title{reverse-compatibility wrapper around compute_pair_deScore()}
\usage{
de_pair(...)
}
\description{
reverse-compatibility wrapper around compute_pair_deScore()
}
\seealso{
\code{link{compute_pair_deScore}}
}
