% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_function_names.R
\name{de_score}
\alias{de_score}
\title{reverse-compatibility wrapper around de_all_pairs_stats()}
\usage{
de_score(...)
}
\arguments{
\item{...}{parameters passed to \code{link{de_all_pairs_stats}}}
}
\description{
reverse-compatibility wrapper around de_all_pairs_stats()
}
\seealso{
\code{link{de_all_pairs_stats}}
}
