% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{de_stats_all_pairs}
\alias{de_stats_all_pairs}
\title{Compute differential expression summary statistics for all pairs of clusters based on de_param()}
\usage{
de_stats_all_pairs(
  norm.dat,
  cl,
  de.param = de_param(),
  method = "limma",
  de.genes = NULL,
  ...
)
}
\arguments{
\item{norm.dat}{a normalized data matrix for data.}

\item{cl}{a cluster factor object.}

\item{de.param}{A list of differential gene expression parameters from \code{de_param()}}

\item{method}{If de.df is NULL, use "limma" or "chisq" to compute differentially expressed genes.}

\item{...}{Additional parameters passed to \code{de_selected_pairs()}}

\item{de.df}{Optional. Pre-computed results from \code{de_all_pairs()} or \code{de_selected_pairs}. Default = NULL.}
}
\value{
a character vector of all differentially expressed genes.
}
\description{
Compute differential expression summary statistics for all pairs of clusters based on de_param()
}
