% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_function_names.R
\name{findVG}
\alias{findVG}
\title{reverse-compatibility wrapper around find_vg()}
\usage{
findVG(...)
}
\description{
reverse-compatibility wrapper around find_vg()
}
\seealso{
\code{link{find_vg}}
}
