% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_functions.R
\name{group_cl}
\alias{group_cl}
\title{Title}
\usage{
group_cl(anno, cluster, norm.dat, select.cells = colnames(norm.dat),
  neun.thresh = 0.5, outlier.cl = NULL, donor.cl = NULL)
}
\arguments{
\item{anno}{anno dataframe with the following required columns - facs_population_plan", "cluster"}

\item{cluster}{clusters labels for all cells along with sample_id as their names}

\item{norm.dat}{expression dataframe with columns as cells and rows as gene names and cpm normalized}

\item{select.cells}{column nmaes of norm.dat}

\item{neun.thresh}{default 0.5}

\item{outlier.cl}{default is NULL}

\item{donor.cl}{default is NULL}
}
\value{
gives the clusters names in each class (exc, inh, etc. ) as a list
}
\description{
Title
}
\examples{
group_cl()
}
