% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_impute.R
\name{impute_knn}
\alias{impute_knn}
\title{Title}
\usage{
impute_knn(knn.idx, reference, dat, knn.dist = NULL, w = NULL, ...)
}
\arguments{
\item{knn.idx}{}

\item{reference}{}

\item{dat}{}

\item{knn.dist}{}

\item{w}{}

\item{...}{}
}
\description{
Title
}
