% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{lm_normalize}
\alias{lm_normalize}
\title{Normalize using a linear model}
\usage{
lm_normalize(dat, x, R_2.th = 0.2, padj.th = 0.01, min.genes = 5)
}
\arguments{
\item{dat}{a data matrix}

\item{x}{a vector}

\item{R_2.th}{R^2 threshold. Default is 0.2.}

\item{padj.th}{p-value threshold. Default is 0.01.}

\item{min.genes}{Minimum genes. Default is 5.}
}
\description{
Normalize using a linear model
}
