% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceDimension_PCA.R
\name{rd_PCA}
\alias{rd_PCA}
\title{Title}
\usage{
rd_PCA(
  norm.dat,
  select.genes = row.names(norm.dat),
  select.cells = colnames(norm.dat),
  sampled.cells = select.cells,
  max.pca = 10,
  th = 2
)
}
\arguments{
\item{norm.dat}{}

\item{select.genes}{}

\item{select.cells}{}

\item{sampled.cells}{}

\item{max.pca}{}

\item{th}{}
}
\value{

}
\description{
Title
}
