% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_consensus.R
\name{run_consensus_clust}
\alias{run_consensus_clust}
\title{Wrapper function to repeatively run clustering on subsampled cells and infer consensus clusters}
\usage{
run_consensus_clust(norm.dat, select.cells = colnames(norm.dat),
  niter = 100, sample.frac = 0.8, de.param = de_param(),
  output_dir = "subsample_result", mc.cores = 1, override = FALSE,
  init.result = NULL, cut.method = "auto", ...)
}
\arguments{
\item{norm.dat}{normalized expression data matrix in log transform, using genes as rows, and cells and columns. Users can use log2(FPKM+1) or log2(CPM+1).}

\item{select.cells}{The cells to be clustered. Default: columns of norm.dat}

\item{niter}{The number of iteractions to run. Default 100.}

\item{sample.frac}{The fraction of of cells sampled per run. Default: 0.8.}

\item{de.param}{The differential gene expression threshold. See de_param() function for details.}

\item{output_dir}{The output directory to store clutering results for each iteraction.}

\item{mc.cores}{The number of cores to be used for parallel processing.}

\item{override}{binary variable determine if the clustering results already stored in output_dir should be overriden.}

\item{init.result}{The pre-set high level clusters. If set, the function will only find finer splits of the current clusters.}

\item{...}{Other parameters passed to iter_clust}
}
\description{
Wrapper function to repeatively run clustering on subsampled cells and infer consensus clusters
}
