% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.genes.R
\name{score_pair_chisq}
\alias{score_pair_chisq}
\title{Perform pairwise differential detection tests using Chi-Squared for a single pair of clusters}
\usage{
score_pair_chisq(pair, cl.present, cl.means, cl.size, genes)
}
\arguments{
\item{pair}{a numeric vector of length 2 specifying which clusters to compare}

\item{cl.present}{a data.frame of gene detection proportions (genes x clusters)}

\item{cl.means}{a data.frame of mean gene expression values (genes x clusters)}

\item{cl.size}{a named numeric vector of cluster sizes}

\item{genes}{the genes to use for pairwise comparisons}
}
\value{
a data.frame with DE statistics:
\itemize{
\item{padj}
\item{pval}
\item{lfc}
\item{meanA}
\item{meanB}
\item{q1}
\item{q2}
}
}
\description{
Perform pairwise differential detection tests using Chi-Squared for a single pair of clusters
}
