% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{select_markers}
\alias{select_markers}
\title{Title}
\usage{
select_markers(norm.dat, cl, n.markers = 20, de.genes = NULL, ...)
}
\arguments{
\item{norm.dat}{}

\item{cl}{}

\item{n.markers}{}

\item{de.genes}{}

\item{...}{}
}
\value{

}
\description{
Title
}
